% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_dmy}
\alias{as_dmy}
\title{Convert input in day-month-year format to date}
\usage{
as_dmy(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Date variable of class Date
}
\description{
Convert input in day-month-year format to date
}
\examples{
as_dmy("1-2-2014")

}

