% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_dmy_hms}
\alias{as_dmy_hms}
\title{Convert input in day-month-year-hour-minute-second format to date-time}
\usage{
as_dmy_hms(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Date-time variable of class Date
}
\description{
Convert input in day-month-year-hour-minute-second format to date-time
}
\examples{
as_mdy_hms("1-1-2014 12:15:01")
}

