% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_duration}
\alias{as_duration}
\title{Wrapper for lubridate's as.duration function. Result converted to numeric}
\usage{
as_duration(x)
}
\arguments{
\item{x}{Time difference}
}
\description{
Wrapper for lubridate's as.duration function. Result converted to numeric
}

