% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_hm}
\alias{as_hm}
\title{Convert input in hour-minute format to time}
\usage{
as_hm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Time variable of class Period
}
\description{
Convert input in hour-minute format to time
}
\examples{
as_hm("12:45")
\dontrun{
as_hm("12:45") \%>\% minute
}
}

