% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_hms}
\alias{as_hms}
\title{Convert input in hour-minute-second format to time}
\usage{
as_hms(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Time variable of class Period
}
\description{
Convert input in hour-minute-second format to time
}
\examples{
as_hms("12:45:00")
\dontrun{
as_hms("12:45:00") \%>\% hour
as_hms("12:45:00") \%>\% second
}
}

