% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_mdy}
\alias{as_mdy}
\title{Convert input in month-day-year format to date}
\usage{
as_mdy(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Date variable of class Date
}
\description{
Convert input in month-day-year format to date
}
\details{
Use as.character if x is a factor
}
\examples{
as_mdy("2-1-2014")
\dontrun{
as_mdy("2-1-2014") \%>\% month(label = TRUE)
as_mdy("2-1-2014") \%>\% week
as_mdy("2-1-2014") \%>\% wday(label = TRUE)
}
}

