% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_mdy_hm}
\alias{as_mdy_hm}
\title{Convert input in month-day-year-hour-minute format to date-time}
\usage{
as_mdy_hm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Date-time variable of class Date
}
\description{
Convert input in month-day-year-hour-minute format to date-time
}
\examples{
as_mdy_hm("1-1-2014 12:15")
}

