% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_numeric}
\alias{as_numeric}
\title{Convert variable to numeric avoiding potential issues with factors}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Numeric
}
\description{
Convert variable to numeric avoiding potential issues with factors
}
\examples{
as_numeric(rnorm(10))
as_numeric(letters)
as_numeric(5:10 \%>\% as.factor)
as.numeric(5:10 \%>\% as.factor)
as_numeric(c("1","2"))

}

