% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_ymd}
\alias{as_ymd}
\title{Convert input in year-month-day format to date}
\usage{
as_ymd(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Date variable of class Date
}
\description{
Convert input in year-month-day format to date
}
\examples{
as_ymd("2013-1-1")

}

