% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_ymd_hm}
\alias{as_ymd_hm}
\title{Convert input in year-month-day-hour-minute format to date-time}
\usage{
as_ymd_hm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Date-time variable of class Date
}
\description{
Convert input in year-month-day-hour-minute format to date-time
}
\examples{
as_ymd_hm("2014-1-1 12:15")
}

