% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_ymd_hms}
\alias{as_ymd_hms}
\title{Convert input in year-month-day-hour-minute-second format to date-time}
\usage{
as_ymd_hms(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Date-time variable of class Date
}
\description{
Convert input in year-month-day-hour-minute-second format to date-time
}
\examples{
as_ymd_hms("2014-1-1 12:15:01")
\dontrun{
as_ymd_hms("2014-1-1 12:15:01") \%>\% as.Date
as_ymd_hms("2014-1-1 12:15:01") \%>\% month
as_ymd_hms("2014-1-1 12:15:01") \%>\% hour
}
}

