% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{auc}
\alias{auc}
\title{Area Under the Curve (AUC)}
\usage{
auc(pred, rvar, lev)
}
\arguments{
\item{pred}{Prediction or predictor}

\item{rvar}{Response variable}

\item{lev}{The level in the response variable defined as _success_}
}
\value{
AUC statistic
}
\description{
Area Under the Curve (AUC)
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/performance.html} for an example in Radiant
}
\examples{
auc(mtcars$mpg, mtcars$vs, 1)

}
\seealso{
\code{\link{performance}} to calculate results

\code{\link{summary.performance}} to summarize results

\code{\link{plot.performance}} to plot results
}

