% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{avengers}
\alias{avengers}
\title{Avengers}
\format{A data frame with 7 rows and 4 variables}
\usage{
data(avengers)
}
\description{
Avengers
}
\details{
List of avengers. The dataset is used to illustrate data merging / joining. Description provided in attr(avengers,"description")
}
\keyword{datasets}

