% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{ci_label}
\alias{ci_label}
\title{Labels for confidence intervals}
\usage{
ci_label(alt = "two.sided", cl = 0.95)
}
\arguments{
\item{alt}{Type of hypothesis ("two.sided","less","greater")}

\item{cl}{Confidence level}
}
\value{
A character vector with labels for a confidence interval
}
\description{
Labels for confidence intervals
}
\examples{
ci_label("less",.95)
ci_label("two.sided",.95)
ci_label("greater",.9)

}

