% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{ci_perc}
\alias{ci_perc}
\title{Values at confidence levels}
\usage{
ci_perc(dat, alt, cl)
}
\arguments{
\item{dat}{Data}

\item{alt}{Type of hypothesis ("two.sided","less","greater")}

\item{cl}{Confidence level}
}
\value{
A charater vector with labels for a confidence interval
}
\description{
Values at confidence levels
}
\examples{
ci_perc(0:100, "less",.95)
ci_perc(0:100, "greater",.95)
ci_perc(0:100, "two.sided",.80)
}

