% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_means.R
\name{compare_means}
\alias{compare_means}
\title{Compare means for two or more variables}
\usage{
compare_means(dataset, cm_var1, cm_var2, cm_paired = "independent",
  cm_alternative = "two.sided", cm_sig_level = 0.95, cm_adjust = "none",
  cm_plots = "bar", cm_jitter = FALSE)
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in a values list from Radiant}

\item{cm_var1}{A numeric variable or factor selected for comparison}

\item{cm_var2}{One or more numeric variable for comparison. If cm_var1 is a factor only one variable can be selected and the mean of this variable is compared across the factor levels of cm_var1}

\item{cm_alternative}{The alternative hypothesis (two.sided, greater or less)}

\item{cm_sig_level}{Span of the confidence interval}

\item{cm_adjust}{Adjustment for multiple comparisons (none or bonferroni)}

\item{cm_plots}{One or more plots of mean values (bar, box, or density plot)}

\item{cm_jitter}{Add jitter to data points for a box plot}
}
\value{
A list with all variables defined in the function
}
\description{
Compare means for two or more variables
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
compare_means("diamonds","price",c("carats","table"))
}
\seealso{
\code{\link{summary_compare_means}} to print results

\code{\link{plots_compare_means}} to plot results
}

