% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{computer}
\alias{computer}
\title{Perceptions of computer (re)sellers
The dataset is used to illustrate perceptual maps
Description provided in attr(computer,"description")}
\format{A data frame with 5 rows and 8 variables}
\usage{
data(computer)
}
\description{
Perceptions of computer (re)sellers
The dataset is used to illustrate perceptual maps
Description provided in attr(computer,"description")
}
\keyword{datasets}

