% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{confint_robust}
\alias{confint_robust}
\title{Confidence interval for robust estimators}
\usage{
confint_robust(object, parm, level = 0.95, vcov = NULL, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{parm}{A specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered}

\item{level}{The confidence level required}

\item{vcov}{Covariance matrix generated by, e.g., sandwich::vcovHC}

\item{...}{Additional argument(s) for methods}
}
\description{
Confidence interval for robust estimators
}
\details{
Wrapper for confint.default with robust standard errors. See \url{http://stackoverflow.com/a/3820125/1974918}
}

