% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint.R
\name{conjoint}
\alias{conjoint}
\title{Conjoint analysis}
\usage{
conjoint(dataset, ca_var1, ca_var2, data_filter = "", show_filter = FALSE,
  ca_rev = FALSE, ca_vif = FALSE, ca_plots = "pw",
  ca_scale_plot = FALSE)
}
\description{
Conjoint analysis
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/conjoint.html} for an example in Radiant
}

