% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conjoint_profiles.R
\name{conjoint_profiles}
\alias{conjoint_profiles}
\title{Create fractional factorial design for conjoint analysis}
\usage{
conjoint_profiles(dataset)
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}
}
\value{
A list with all variables defined in the function as an object of class conjoint_profiles
}
\description{
Create fractional factorial design for conjoint analysis
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/conjoint_profiles.html} for an example in Radiant
}
\examples{
\donttest{
 cp <<- c("price = c('$10','$13','$16')", "sight = c('Staggered','Not Staggered')",
        "comfort = c('Average no cupholder','Average cupholder','Large cupholder')",
        "audio.visual = c('Small plain','Large plain','Large digital')",
        "food = c('No food','Hot dogs and popcorn','Gourmet food')")
 result <- conjoint_profiles("cp")
 result <- cp \%>\% conjoint_profiles
 rm(cp, envir = .GlobalEnv)
}
}
\seealso{
\code{\link{summary.conjoint_profiles}} to summarize results
}

