% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{copy_from}
\alias{copy_from}
\title{Source for package functions}
\usage{
copy_from(.from, ...)
}
\arguments{
\item{.from}{The package to pull the function from}

\item{...}{Functions to pull}
}
\description{
Source for package functions
}
\details{
Equivalent of source with local=TRUE for package functions. Written by smbache, author of the import package. See \url{https://github.com/smbache/import/issues/4} for a discussion. This function will be depracated when (if) it is included in \url{https://github.com/smbache/import}
}
\examples{

copy_from(radiant, getdata)

}

