% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{copy_imported}
\alias{copy_imported}
\title{Import all functions that a package imports for use with Shiny}
\usage{
copy_imported(.from)
}
\arguments{
\item{.from}{The package to pull the function from}
}
\description{
Import all functions that a package imports for use with Shiny
}
\examples{
\dontrun{
copy_imported(radiant)
}
}

