% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{cv}
\alias{cv}
\title{Coefficient of variation}
\usage{
cv(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Coefficient of variation
}
\description{
Coefficient of variation
}
\examples{
cv(runif (100))

}

