% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{dfprint}
\alias{dfprint}
\title{Print a data.frame with a specified number of decimal places}
\usage{
dfprint(tbl, dec = 3, perc = FALSE)
}
\arguments{
\item{tbl}{Data.frame}

\item{dec}{Number of decimal places}

\item{perc}{Display numbers as percentages (TRUE or FALSE)}
}
\value{
Data.frame for printing
}
\description{
Print a data.frame with a specified number of decimal places
}
\examples{
data.frame(x = c("a","b"), y = c(1L, 2L), z = c(-0.0005, 3)) \%>\%
  dfprint(dec = 3)

}

