% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{dfround}
\alias{dfround}
\title{Round double in a data.frame to a specified number of decimal places}
\usage{
dfround(tbl, dec = 3)
}
\arguments{
\item{tbl}{Data.frame}

\item{dec}{Number of decimal places}
}
\value{
Data.frame for viewing
}
\description{
Round double in a data.frame to a specified number of decimal places
}
\examples{
data.frame(x = c("a","b"), y = c(1L, 2L), z = c(-0.0005, 3.1)) \%>\%
  dfround(dec = 3)

}

