% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{diamonds}
\alias{diamonds}
\title{A sample of 3,000 from the diamonds dataset bundeled with ggplot2
Description provided in attr(diamonds,"description")}
\format{A data frame with 3000 rows and 10 variables}
\usage{
data(diamonds)
}
\description{
A sample of 3,000 from the diamonds dataset bundeled with ggplot2
Description provided in attr(diamonds,"description")
}
\keyword{datasets}

