% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{do_filter}
\alias{do_filter}
\title{Workaround for https://github.com/rstudio/shiny/issues/879}
\usage{
do_filter(dat, sel)
}
\arguments{
\item{dat}{Data}

\item{sel}{Filter string}
}
\description{
Workaround for https://github.com/rstudio/shiny/issues/879
}

