% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe.R
\name{doe}
\alias{doe}
\title{Create (partial) factorial design}
\usage{
doe(factors, int = "", trials = NA, seed = NA)
}
\arguments{
\item{factors}{Categorical variables used as input for design}

\item{int}{Vector of interaction terms to consider when generating design}

\item{trials}{Number of trial to create. If NA then all feasible designs will be considered until a design with perfect D-efficiency is found}

\item{seed}{Random seed to use as the starting point}
}
\value{
A list with all variables defined in the function as an object of class conjoint_profiles
}
\description{
Create (partial) factorial design
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/doe.html} for an example in Radiant
}
\examples{
"price; $10; $13; $16\\nfood; popcorn; gourmet; no food" \%>\% doe

}
\seealso{
\code{\link{summary.conjoint_profiles}} to summarize results
}

