% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtree}
\alias{dtree}
\title{Create a decision tree}
\usage{
dtree(yl)
}
\arguments{
\item{yl}{A yaml string or a list (e.g., from yaml::yaml.load_file())}
}
\value{
A list with the initial tree and the calculated tree
}
\description{
Create a decision tree
}
\details{
See \url{http://vnijs.github.io/radiant/base/dtree.html} for an example in Radiant
}
\seealso{
\code{\link{summary.dtree}} to summarize results

\code{\link{plot.dtree}} to plot results
}

