% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtree_parser}
\alias{dtree_parser}
\title{Parse yaml input for dtree to provide (more) useful error messages}
\usage{
dtree_parser(yl)
}
\arguments{
\item{yl}{A yaml string}
}
\value{
An updated yaml string or a vector messages to return to the users
}
\description{
Parse yaml input for dtree to provide (more) useful error messages
}
\details{
See \url{http://vnijs.github.io/radiant/base/dtree.html} for an example in Radiant
}
\seealso{
\code{\link{dtree}} to calculate tree

\code{\link{summary.dtree}} to summarize results

\code{\link{plot.dtree}} to plot results
}

