% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint_profiles.R
\name{ff_design}
\alias{ff_design}
\title{Function to generate a fractional factorial design}
\usage{
ff_design(attr, trial = 0, rseed = 172110)
}
\arguments{
\item{attr}{Attributes used to generate profiles}

\item{trial}{Number of trials that have already been run}

\item{rseed}{Random seed to use}
}
\description{
Function to generate a fractional factorial design
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/conjoint_profiles.html} for an example in Radiant
}
\seealso{
\code{\link{conjoint_profiles}} to calculate results

\code{\link{summary.conjoint_profiles}} to summarize results
}

