% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{filter_na}
\alias{filter_na}
\title{Filter out rows with missing observations}
\usage{
filter_na(data, vars = "")
}
\arguments{
\item{data}{Data.frame to filter}

\item{vars}{Variables used to identify rows with missing values (default is all)}
}
\description{
Filter out rows with missing observations
}
\details{
Filter data on complete cases for selected variables
}
\examples{
mtcars[1:4, "mpg"] <- NA; mtcars[5:10, "carb"] <- NA
filter_na(mtcars, "mpg")
}

