% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{filterdata}
\alias{filterdata}
\title{Filter data with user-specified expression}
\usage{
filterdata(dat, filt = "")
}
\arguments{
\item{dat}{Data.frame to filter}

\item{filt}{Filter expression to apply to the specified dataset (e.g., "price > 10000" if dataset is "diamonds")}
}
\value{
Filtered data.frame
}
\description{
Filter data with user-specified expression
}

