% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{find_dropbox}
\alias{find_dropbox}
\title{Find a users dropbox directory}
\usage{
find_dropbox(folder = 1)
}
\arguments{
\item{folder}{If multiple folders are present select which one to use. The first folder listed is used by default.}
}
\value{
Path to users personal dropbox directory
}
\description{
Find a users dropbox directory
}

