% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{find_max}
\alias{find_max}
\title{Find maxium value of a vector}
\usage{
find_max(var, val = "")
}
\arguments{
\item{var}{Variable to find the maximum for}

\item{val}{Variable to find the value for at the maxium of var}
}
\value{
Value of val at the maximum of var
}
\description{
Find maxium value of a vector
}

