% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{find_min}
\alias{find_min}
\title{Find minimum value of a vector}
\usage{
find_min(var, val = "")
}
\arguments{
\item{var}{Variable to find the minimum for}

\item{val}{Variable to find the value for at the maxium of var}
}
\value{
Value of val at the minimum of var
}
\description{
Find minimum value of a vector
}

