% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{flip}
\alias{flip}
\title{Flip the DT table to put Function, Variable, or Group by on top}
\usage{
flip(expl, top = "fun")
}
\arguments{
\item{expl}{Return value from \code{\link{explore}}}

\item{top}{The variable (type) to display at the top of the table ("fun" for Function, "var" for Variable, and "byvar" for Group by. "fun" is the default}
}
\description{
Flip the DT table to put Function, Variable, or Group by on top
}
\details{
See \url{http://vnijs.github.io/radiant/base/explore.html} for an example in Radiant
}
\examples{
result <- explore("diamonds", "price:x") \%>\% flip("var")

result <- explore("diamonds", "price", byvar = "cut", fun = c("length", "skew")) \%>\%
  flip("byvar")

}
\seealso{
\code{\link{explore}} to generate summaries

\code{\link{make_expl}} to create the DT table
}

