% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{getclass}
\alias{getclass}
\title{Get variable class}
\usage{
getclass(dat)
}
\arguments{
\item{dat}{Dataset to evaluate}
}
\value{
Vector with class information for each variable
}
\description{
Get variable class
}
\details{
Get variable class information for each column in a data.frame
}
\examples{
getclass(mtcars)

}

