% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{getdata_exp}
\alias{getdata_exp}
\title{Get data for analysis functions exported by Radiant}
\usage{
getdata_exp(dataset, vars = "", na.rm = TRUE, filt = "")
}
\arguments{
\item{dataset}{Name of the dataframe}

\item{vars}{Variables to extract from the dataframe}

\item{na.rm}{Remove rows with missing values (default is TRUE)}

\item{filt}{Filter to apply to the specified dataset. For example "price > 10000" if dataset is "diamonds" (default is "")}
}
\value{
Data.frame with the specified columns selected
}
\description{
Get data for analysis functions exported by Radiant
}

