% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{is_not}
\alias{is_not}
\title{Convenience function for is.null or is.na}
\usage{
is_not(x)
}
\arguments{
\item{x}{Input}
}
\description{
Convenience function for is.null or is.na
}
\examples{
is_not(NA)
is_not(NULL)

}

