% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{is_string}
\alias{is_string}
\title{Is input a string?}
\usage{
is_string(x)
}
\arguments{
\item{x}{Input}
}
\value{
TRUE if string, else FALSE
}
\description{
Is input a string?
}
\details{
Is input a string
}
\examples{
is_string("")
is_string("data")
is_string(c("data","data"))
is_string(NULL)

}

