% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{iterms}
\alias{iterms}
\title{Create a vector of interaction terms}
\usage{
iterms(vars, nway, sep = ":")
}
\arguments{
\item{vars}{Variables lables to use}

\item{nway}{2-way (2) or 3-way (3) interactions labels to create}

\item{sep}{Separator between variable names (default is :)}
}
\value{
Character vector of interaction term labels
}
\description{
Create a vector of interaction terms
}
\examples{
paste0("var", 1:3) \%>\% iterms(2)
paste0("var", 1:3) \%>\% iterms(3)
paste0("var", 1:3) \%>\% iterms(2, sep = ".")

}

