% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{level_list}
\alias{level_list}
\title{Generate list of levels and unique values}
\usage{
level_list(dat, ...)
}
\arguments{
\item{dat}{A data.frame}

\item{...}{Unquoted variable names to evaluate}
}
\description{
Generate list of levels and unique values
}
\examples{
data.frame(a = c(rep("a",5),rep("b",5)), b = c(rep(1,5),6:10)) \%>\% level_list
level_list(mtcars, mpg, cyl)

}

