% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{lin_launcher}
\alias{lin_launcher}
\title{Create a launcher and updater for Linux (.sh)}
\usage{
lin_launcher(app = c("analytics", "marketing", "quant", "base"))
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("analytics", "marketing", "quant", or "base"). Default is "analytics"}
}
\description{
Create a launcher and updater for Linux (.sh)
}
\details{
On Linux a file named 'radiant.sh' and one named 'update_radiant.sh' will be put on the desktop. Double-click the file to launch the specified Radiant app or update Radiant to the latest version
}
\examples{
if (interactive()) {
  if (Sys.info()["sysname"] == "Linux") {
    lin_launcher()
    fn <- paste0("/home/",Sys.getenv("USER"),"/Desktop/radiant.sh")
    if (!file.exists(fn))
      stop("Linux launcher not created")
    else
      unlink(fn)
  }
}
}

