% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{ln}
\alias{ln}
\title{Natural log}
\usage{
ln(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{Remove missing values (default is TRUE)}
}
\value{
Natural log of vector
}
\description{
Natural log
}
\examples{
ln(runif(10,1,2))

}

