% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{loadcsv}
\alias{loadcsv}
\title{Load a csv files with read.csv and read_csv}
\usage{
loadcsv(fn, header = TRUE, sep = ",", saf = TRUE, safx = 10)
}
\arguments{
\item{fn}{File name string}

\item{header}{Header in file (TRUE, FALSE)}

\item{sep}{Use , or ; or \\t}

\item{saf}{Convert character variables to factors if (1) there are less than 100 distinct values (2) there are X (see safx) more values than levels}

\item{safx}{Values to levels ratio}
}
\value{
Data.frame with (some) variables converted to factors
}
\description{
Load a csv files with read.csv and read_csv
}

