% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{loadrda}
\alias{loadrda}
\title{Load an rda file and add it to the radiant data list (r_data)}
\usage{
loadrda(fn, ext = "rda")
}
\arguments{
\item{fn}{File name string}

\item{ext}{File extension ("rda" is the default)}
}
\value{
Data.frame in r_data
}
\description{
Load an rda file and add it to the radiant data list (r_data)
}

