% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{loadrda_url}
\alias{loadrda_url}
\title{Load an rda file from a url}
\usage{
loadrda_url(rda_url)
}
\arguments{
\item{rda_url}{URL for the csv file}
}
\value{
Data.frame
}
\description{
Load an rda file from a url
}

