% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore.R
\name{make_expl}
\alias{make_expl}
\title{Make a tabel of summary statistics in DT}
\usage{
make_expl(expl, format = "none")
}
\arguments{
\item{expl}{Return value from \code{\link{explorer}}}

\item{format}{Show Color bar ("color_bar"),  Heat map ("heat"), or None ("none")}
}
\description{
Make a tabel of summary statistics in DT
}
\details{
See \url{http://vnijs.github.io/radiant/base/explore.html} for an example in Radiant
}
\examples{
pivotr("diamonds", cvars = "cut") \%>\% make_dt
pivotr("diamonds", cvars = c("cut","clarity")) \%>\% make_dt(format = "color_bar")
ret <-  pivotr("diamonds", cvars = c("cut","clarity"), normalize = "total") \%>\%
   make_dt(format = "color_bar", check = "perc")
}
\seealso{
\code{\link{pivotr}} to create the pivot-table using dplyr

\code{\link{summary.pivotr}} to print a plain text table
}

