% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{make_funs}
\alias{make_funs}
\title{Make a list of functions-as-formulas to pass to dplyr}
\usage{
make_funs(x)
}
\arguments{
\item{x}{List of functions as strings}
}
\value{
List of functions to pass to dplyr in formula form
}
\description{
Make a list of functions-as-formulas to pass to dplyr
}
\examples{
make_funs(c("mean", "sum_rm"))

}

