% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{max_rm}
\alias{max_rm}
\title{Max with na.rm = TRUE}
\usage{
max_rm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Maximum value
}
\description{
Max with na.rm = TRUE
}
\examples{
max_rm(runif (100))

}

