% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mean_rm}
\alias{mean_rm}
\title{Mean with na.rm = TRUE}
\usage{
mean_rm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Mean value
}
\description{
Mean with na.rm = TRUE
}
\examples{
mean_rm(runif (100))

}

