% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mode_rm}
\alias{mode_rm}
\title{Mode with na.rm = TRUE}
\usage{
mode_rm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Mode value
}
\description{
Mode with na.rm = TRUE
}
\examples{
mode_rm(diamonds$cut)

}

