% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{mp3}
\alias{mp3}
\title{Conjoint data for MP3 players
Description provided in attr(mp3,"description")}
\format{A data frame with 18 rows and 6 variables}
\usage{
data(mp3)
}
\description{
Conjoint data for MP3 players
Description provided in attr(mp3,"description")
}
\keyword{datasets}

