% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{mutate_each_add}
\alias{mutate_each_add}
\title{Add tranformed variables to a data.frame (NSE)}
\usage{
mutate_each_add(data, fun, ext = "", ...)
}
\arguments{
\item{data}{Data.frame to add transformed variables to}

\item{fun}{Function to apple (e.g., "log")}

\item{ext}{Extension to add for each variable}

\item{...}{Variables to transform}
}
\description{
Add tranformed variables to a data.frame (NSE)
}
\examples{
mutate_each_add(mtcars, fun = "log", ext = "_log", mpg, cyl)
}

