% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{mutate_each_add_}
\alias{mutate_each_add_}
\title{Add tranformed variables to a data.frame (SE)}
\usage{
mutate_each_add_(data, fun, ext = "", vars = "")
}
\arguments{
\item{data}{Data.frame to add transformed variables to}

\item{fun}{Function to apple (e.g., "log")}

\item{ext}{Extension to add for each variable}

\item{vars}{Variables to transform (default is all)}
}
\description{
Add tranformed variables to a data.frame (SE)
}
\examples{
mutate_each_add_(mtcars, fun = "log", ext = "_log", vars = c("mpg","cyl"))
}

