% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{n_missing}
\alias{n_missing}
\title{Number of missing values}
\usage{
n_missing(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
number of missing values
}
\description{
Number of missing values
}
\examples{
n_missing(c("a","b",NA))

}

