% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore.R
\name{nmissing}
\alias{nmissing}
\title{Number of missing values}
\usage{
nmissing(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
number of missing values
}
\description{
Number of missing values
}
\examples{
nmissing(c("a","b",NA))
}

