% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize}
\alias{normalize}
\title{Normalize a variable x by a variable y}
\usage{
normalize(x, y)
}
\arguments{
\item{x}{Input variable}

\item{y}{Normalizing variable}
}
\value{
x/y
}
\description{
Normalize a variable x by a variable y
}

