% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize_each_add}
\alias{normalize_each_add}
\title{Add normalized variables to a data.frame (NSE)}
\usage{
normalize_each_add(data, ext = "", ...)
}
\arguments{
\item{data}{Data.frame to add transformed variables to}

\item{ext}{Extension to add for each variable}

\item{...}{Variables to normalize. The first variable is used as the normalizer}
}
\description{
Add normalized variables to a data.frame (NSE)
}
\examples{
normalize_each_add(mtcars, ext = "_nz", mpg, cyl, mpg)
}

