% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{ntile}
\alias{ntile}
\title{Wrapper for dplyr::ntile}
\usage{
ntile(x, n)
}
\arguments{
\item{x}{Numeric variable}

\item{n}{number of bins to create}
}
\description{
Wrapper for dplyr::ntile
}
\details{
Wrapper for dplyr::ntile to ensure result is an integer
}
\examples{
ntile(1:10,5)
}

