% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{outer_join}
\alias{outer_join}
\title{Outer join}
\usage{
outer_join(dataset, dataset2, by = intersect(names(dataset), names(dataset2)))
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{dataset2}{Dataset name (string) to merge with `dataset`. This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{by}{Variables used to merge/join `dataset` and `dataset2`}
}
\value{
The merged data.frame
}
\description{
Outer join
}
\details{
Join two data.frames, keeping rows and columns that appear in either
}
\examples{
outer_join(mtcars[,-1], mtcars[,-5])
outer_join(mtcars[,1:3], mtcars[,2:4])
outer_join(mtcars[,1:3], mtcars[,2:4], by = c("cyl","disp"))
}

