% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p05}
\alias{p05}
\title{5th percentile}
\usage{
p05(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
5th percentile
}
\description{
5th percentile
}
\examples{
p05(rnorm(100))

}

