% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p10}
\alias{p10}
\title{10th percentile}
\usage{
p10(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
10th percentile
}
\description{
10th percentile
}
\examples{
p10(rnorm(100))

}

