% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p25}
\alias{p25}
\title{25th percentile}
\usage{
p25(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
25th percentile
}
\description{
25th percentile
}
\examples{
p25(rnorm(100))

}

