% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p75}
\alias{p75}
\title{75th percentile}
\usage{
p75(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
75th percentile
}
\description{
75th percentile
}
\examples{
p75(rnorm(100))

}

