% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p90}
\alias{p90}
\title{90th percentile}
\usage{
p90(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
90th percentile
}
\description{
90th percentile
}
\examples{
p90(rnorm(100))

}

