% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p95}
\alias{p95}
\title{95th percentile}
\usage{
p95(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
95th percentile
}
\description{
95th percentile
}
\examples{
p95(rnorm(100))

}

