% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Model performance}
\usage{
performance(dataset, pred, rvar, lev = "", qnt = 10, method = "xtile",
  train = "", data_filter = "")
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{pred}{Prediction or predictor}

\item{rvar}{Response variable}

\item{lev}{The level in the response variable defined as _success_}

\item{qnt}{Number of bins to create}

\item{method}{Use either ntile or xtile to split the data (default is xtile)}

\item{train}{Use data from training ("train"), validation ("valid"), or both ("both") to evaluate model performance}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}
}
\value{
A list of results
}
\description{
Model performance
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/performance.html} for an example in Radiant
}
\examples{
result <- performance("titanic", c("age","fare"), "survived")

}
\seealso{
\code{\link{summary.performance}} to summarize results

\code{\link{plot.performance}} to plot results
}

