% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pivotr.R
\name{pivotr}
\alias{pivotr}
\title{Create a pivot table using dplyr}
\usage{
pivotr(dataset, cvars = "", nvar = "None", fun = "mean",
  normalize = "None", data_filter = "", shiny = FALSE)
}
\arguments{
\item{dataset}{Name of the dataframe to change}

\item{cvars}{Categorical variables}

\item{nvar}{Numerical variable}

\item{fun}{Function to apply to numerical variable}

\item{normalize}{Normalize the table by "row" total,"colum" totals, or overall "total"}

\item{data_filter}{Expression used to filter the dataset. This should be a string (e.g., "price > 10000")}

\item{shiny}{Logical (TRUE, FALSE) to indicate if the function call originate inside a shiny app}
}
\description{
Create a pivot table using dplyr
}
\details{
Create a pivot-table. See \url{http://vnijs.github.io/radiant/base/pivotr.html} for an example in Radiant
}
\examples{
pivotr("diamonds", cvars = "cut")$tab
pivotr("diamonds", cvars = c("cut","clarity","color"))$tab
pivotr("diamonds", cvars = "cut:clarity", nvar = "price")$tab
}

