% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_means.R
\name{plot.compare_means}
\alias{plot.compare_means}
\title{Plot results from the compare_means function. This is a method of class compare_means and can be called as plot or plot.compare_means}
\usage{
\method{plot}{compare_means}(result)
}
\description{
Plot results from the compare_means function. This is a method of class compare_means and can be called as plot or plot.compare_means
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
result <- compare_means("diamonds","price",c("carats","table"))
plot(result)
}
\seealso{
\code{\link{compare_means}} to calculate results

\code{\link{summary.compare_means}} to summarize results
}

