% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_prop.R
\name{plot.compare_props}
\alias{plot.compare_props}
\title{Plot results from the compare_props function. This is a method of class compare_props and can be called as plot or plot.compare_props}
\usage{
\method{plot}{compare_props}(result)
}
\description{
Plot results from the compare_props function. This is a method of class compare_props and can be called as plot or plot.compare_props
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/compare_props.html} for an example in Radiant
}
\examples{
result <- compare_props("titanic", "pclass", "survived", cp_plots = "props")
plot(result)
}
\seealso{
\code{\link{compare_props}} to calculate results

\code{\link{summary.compare_props}} to summarize results
}

