% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot.correlation}
\alias{plot.correlation}
\title{Plot results from the correlation function. This is a method of class correlation and can be called as plot or plot.correlation}
\usage{
\method{plot}{correlation}(result)
}
\description{
Plot results from the correlation function. This is a method of class correlation and can be called as plot or plot.correlation
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/correlation.html} for an example in Radiant
}
\examples{
result <- correlation("titanic", "pclass", "survived", cp_plots = "props")
plot(result)
}
\seealso{
\code{\link{correlation}} to calculate results

\code{\link{summary.correlation}} to summarize results
}

