% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot.correlation_}
\alias{plot.correlation_}
\title{Plot method for the correlation function}
\usage{
\method{plot}{correlation_}(x, ...)
}
\arguments{
\item{x}{Return value from \code{\link{correlation}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for the correlation function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/correlation.html} for an example in Radiant
}
\examples{
result <- correlation("diamonds",c("price","carat","clarity"))
plot(result)
diamonds \%>\% correlation("price:clarity") \%>\% plot

}
\seealso{
\code{\link{correlation}} to calculate results

\code{\link{summary.correlation_}} to summarize results
}

