% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cross_tabs.R
\name{plot.cross_tabs}
\alias{plot.cross_tabs}
\title{Plot method for the cross_tabs function}
\usage{
\method{plot}{cross_tabs}(result, ct_check = "")
}
\description{
Plot method for the cross_tabs function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/cross_tabs.html} for an example in Radiant
}
\examples{
result <- cross_tabs("newspaper", "Income", "Newspaper", ct_expected = TRUE)
plot(result)
}
\seealso{
\code{\link{cross_tabs}} to calculate results

\code{\link{summary.cross_tabs}} to summarize results
}

