% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{plot.crs}
\alias{plot.crs}
\title{Plot method for the crs function}
\usage{
\method{plot}{crs}(x, shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{crs}}}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the crs function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/crs.html} for an example in Radiant
}
\seealso{
\code{\link{crs}} to generate results

\code{\link{summary.crs}} to summarize results
}

