% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/full_factor.R
\name{plot.full_factor}
\alias{plot.full_factor}
\title{Plot method for the full_factor function}
\usage{
\method{plot}{full_factor}(x, ...)
}
\arguments{
\item{x}{Return value from \code{\link{full_factor}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the full_factor function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/full_factor.html} for an example in Radiant
}
\examples{
result <- full_factor("diamonds",c("price","carat","table"))
plot(result)
result <- full_factor("computer","HighEnd:Business")
summary(result)
}
\seealso{
\code{\link{full_factor}} to calculate results

\code{\link{plot.full_factor}} to plot results
}

