% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{plot.kmeans_clus}
\alias{plot.kmeans_clus}
\title{Plot method for kmeans_clus}
\usage{
\method{plot}{kmeans_clus}(x, ...)
}
\arguments{
\item{x}{Return value from \code{\link{kmeans_clus}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for kmeans_clus
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/kmeans_clus.html} for an example in Radiant
}
\examples{
result <- kmeans_clus("shopping", km_vars = c("v1:v6"))
plot(result)
}
\seealso{
\code{\link{kmeans_clus}} to generate results

\code{\link{summary.kmeans_clus}} to summarize results

\code{\link{save_membership}} to add cluster membership to the selected dataset
}

