% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mds.R
\name{plot.mds}
\alias{plot.mds}
\title{Plot method for the mds function}
\usage{
\method{plot}{mds}(x, rev_dim = "", fontsz = 1.3, ...)
}
\arguments{
\item{x}{Return value from \code{\link{mds}}}

\item{rev_dim}{Flip the axes in plots}

\item{fontsz}{Font size to use in plots}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the mds function
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/mds.html} for an example in Radiant
}
\examples{
result <- mds("city","from","to","distance")
plot(result)
plot(result, rev_dim = 1:2)
plot(result, rev_dim = 1:2, fontsz = 2)
}
\seealso{
\code{\link{mds}} to calculate results

\code{\link{summary.mds}} to plot results
}

