% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{plot.performance}
\alias{plot.performance}
\title{Plot method for the performance function}
\usage{
\method{plot}{performance}(x, plots = c("lift", "gains"), shiny = FALSE,
  ...)
}
\arguments{
\item{x}{Return value from \code{\link{performance}}}

\item{plots}{Plots to return}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the performance function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/performance.html} for an example in Radiant
}
\examples{
performance("titanic", "age", "survived") \%>\% plot
performance("titanic", c("age","fare"), "survived") \%>\% plot
performance("titanic", c("age","fare"), "survived", method = "xtile") \%>\% plot
performance("titanic", c("age","fare"), "survived") \%>\% summary
}
\seealso{
\code{\link{performance}} to generate results

\code{\link{summary.performance}} to summarize results
}

