% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_factor.R
\name{plot.pre_factor}
\alias{plot.pre_factor}
\title{Plot method for the pre_factor function}
\usage{
\method{plot}{pre_factor}(x, plots = c("scree", "change"), cutoff = 0.2,
  shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{pre_factor}}}

\item{plots}{Plots to return. "change" shows the change in eigenvalues as variables are grouped into different number of factors, "scree" shows a scree plot of eigenvalues}

\item{cutoff}{For large datasets plots can take time to render and become hard to interpret. By selection a cutoff point (e.g., eigenvalues of .8 or higher) factors with the least explanatory power are removed from the plot}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the pre_factor function
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/pre_factor.html} for an example in Radiant
}
\examples{
result <- pre_factor("diamonds",c("price","carat","table"))
plot(result)
plot(result, plots = c("change", "scree"), cutoff = .05)

}
\seealso{
\code{\link{pre_factor}} to calculate results

\code{\link{summary.pre_factor}} to summarize results
}

