% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_pois}
\alias{plot.prob_pois}
\title{Plot method for the probability calculator function (Poisson distribution)}
\usage{
\method{plot}{prob_pois}(x, type = "values", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_pois}}}

\item{type}{Probabilities or values}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator function (Poisson distribution)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}

