% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_tdist}
\alias{plot.prob_tdist}
\title{Plot method for the probability calculator (t-distribution)}
\usage{
\method{plot}{prob_tdist}(x, type = "values", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_tdist}}}

\item{type}{Probabilities or values}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (t-distribution)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/probability.html} for an example in Radiant
}

