% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_unif}
\alias{plot.prob_unif}
\title{Plot method for the probability calculator (uniform)}
\usage{
\method{plot}{prob_unif}(x, type = "values", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_norm}}}

\item{type}{Probabilities or values}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (uniform)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/probability.html} for an example in Radiant
}

