% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression.R
\name{plot.reg_predict}
\alias{plot.reg_predict}
\title{Plot method for the predict.regression function}
\usage{
\method{plot}{reg_predict}(x, reg_xvar = "", reg_facet_row = ".",
  reg_facet_col = ".", reg_color = "none", reg_conf_level = 0.95, ...)
}
\arguments{
\item{x}{Return value from \code{\link{predict.regression}}.}

\item{reg_xvar}{Variable to display along the X-axis of the plot}

\item{reg_facet_row}{Create vertically arranged subplots for each level of the selected factor variable}

\item{reg_facet_col}{Create horizontally arranged subplots for each level of the selected factor variable}

\item{reg_color}{Adds color to a scatter plot to generate a heat map. For a line plot one line is created for each group and each is assigned a different colour}

\item{reg_conf_level}{Confidence level to use for prediction intervals (.95 is the default). Note that the error bars for predicitions are approximations at this point.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the predict.regression function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
result <- regression("titanic", "survived", c("pclass","sex","age"), reg_levels = "Yes")
pred <- predict(result, reg_predict_cmd = "pclass = levels(pclass)")
plot(pred, reg_xvar = "pclass")
pred <- predict(result, reg_predict_cmd = "age = 0:100")
plot(pred, reg_xvar = "age")
pred <- predict(result, reg_predict_cmd = "pclass = levels(pclass), sex = levels(sex)")
plot(pred, reg_xvar = "pclass", reg_color = "sex")
pred <- predict(result, reg_predict_cmd = "pclass = levels(pclass), age = seq(0,100,20)")
plot(pred, reg_xvar = "pclass", reg_color = "age")
plot(pred, reg_xvar = "age", reg_color = "pclass")
pred <- predict(result, reg_predict_cmd="pclass=levels(pclass), sex=levels(sex), age=seq(0,100,20)")
plot(pred, reg_xvar = "age", reg_color = "sex", reg_facet_col = "pclass")
plot(pred, reg_xvar = "age", reg_color = "pclass", reg_facet_col = "sex")
pred <- predict(result, reg_predict_cmd="pclass=levels(pclass), sex=levels(sex), age=seq(0,100,5)")
plot(pred, reg_xvar = "age", reg_color = "sex", reg_facet_col = "pclass")
plot(pred, reg_xvar = "age", reg_color = "pclass", reg_facet_col = "sex")
}
\seealso{
\code{\link{regression}} to generate the result

\code{\link{summary.regression}} to summarize results

\code{\link{plot.regression}} to plot results

\code{\link{predict.regression}} to generate predictions
}

