% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression.R
\name{plot.regression}
\alias{plot.regression}
\title{Plot results from the regression function. This is a method of class regression and can be called as plot or plot.regression}
\usage{
\method{plot}{regression}(result)
}
\arguments{
\item{result}{Return value from \code{\link{regression}}}
}
\description{
Plot results from the regression function. This is a method of class regression and can be called as plot or plot.regression
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
result <- regression("diamonds", "price", c("carat","clarity"), reg_plots = "dashboard")
plot(result)
}
\seealso{
\code{\link{regression}} to generate the result

\code{\link{summary.regression}} to summarize results
}

