% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulater.R
\name{plot.simulater}
\alias{plot.simulater}
\title{Plot method for the simulater function}
\usage{
\method{plot}{simulater}(x, shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{simulater}}}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the simulater function
}
\details{
See \url{http://vnijs.github.io/radiant/base/simulater} for an example in Radiant
}
\examples{
result <- simulater(const = "cost 3", norm = "demand 2000 1000",
                    discrete = "price 5 .3 8 .7",
                    form = "profit = demand * (price - cost)")
plot(result)
}
\seealso{
\code{\link{single_mean}} to generate the result

\code{\link{summary.single_mean}} to summarize results
}

