% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_prop.R
\name{plot.single_prop}
\alias{plot.single_prop}
\title{Plot method for single_prop}
\usage{
\method{plot}{single_prop}(result)
}
\arguments{
\item{result}{Return value from \code{\link{single_prop}}}
}
\description{
Plot method for single_prop
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_prop.html} for an example in Radiant
}
\examples{
result <- single_prop("diamond","cut")
plot(result)
}
\seealso{
\code{\link{single_prop}} to generate the result

\code{\link{summary.single_prop}} to summarize the results
}

