% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_means.R
\name{plots_compare_means}
\alias{plots_compare_means}
\title{Plot output from compare_means for two or more variables}
\usage{
plots_compare_means(result)
}
\description{
Plot output from compare_means for two or more variables
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
result <- compare_means("diamonds","price",c("carats","table"))
plots_compare_means(result)
}
\seealso{
\code{\link{compare_means}} to calculate results

\code{\link{summary_compare_means}} to plot results
}

