% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_mean.R
\name{plots_single_mean}
\alias{plots_single_mean}
\title{Plot results from the single_mean function}
\usage{
plots_single_mean(result)
}
\arguments{
\item{result}{Return value from \code{\link{single_mean}}}
}
\description{
Plot results from the single_mean function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_mean.html} for an example in Radiant
}
\examples{
result <- single_mean("diamonds","price")
plots_single_mean(result)
}
\seealso{
\code{\link{single_mean}} to generate the result

\code{\link{summary_single_mean}} to print results
}

