% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre_factor.R
\name{pre_factor}
\alias{pre_factor}
\title{Evaluate if data are appropriate for PCA / Factor analysis}
\usage{
pre_factor(dataset, pf_var, data_filter = "")
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{pf_var}{Variables to include in the analysis}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}
}
\value{
A list with all variables defined in the function as an object of class pre_factor
}
\description{
Evaluate if data are appropriate for PCA / Factor analysis
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/pre_factor.html} for an example in Radiant
}
\examples{
result <- pre_factor("diamonds",c("price","carat","table"))
}
\seealso{
\code{\link{summary.pre_factor}} to summarize results

\code{\link{plot.pre_factor}} to plot results
}

