% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ann.R
\name{predict.ann}
\alias{predict.ann}
\title{Predict method for the ann function}
\usage{
\method{predict}{ann}(object, dataset, ...)
}
\arguments{
\item{dataset}{Dataset to use for prediction}

\item{...}{further arguments passed to or from other methods}

\item{x}{Return value from \code{\link{ann}}}
}
\description{
Predict method for the ann function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/ann.html} for an example in Radiant
}
\seealso{
\code{\link{ann}} to generate results

\code{\link{summary.ann}} to generate results

\code{\link{plot.ann}} to plot results
}

