% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation.R
\name{print.correlation}
\alias{print.correlation}
\title{Print method for the correlation function}
\usage{
\method{print}{correlation}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{object}{Return value from \code{\link{correlation}}}
}
\description{
Print method for the correlation function
}
\details{
Default print method throws an error, "Error in if (p > 1) \{ : argument is of length zero". Setting class to 'list' and printing resolves this error
}
\examples{
correlation("diamonds",c("price","carat","clarity"))
}
\seealso{
\code{\link{correlation}} to calculate results

\code{\link{summary.correlation}} to print results

\code{\link{plot.correlation}} to plot results
}

