% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_binom}
\alias{prob_binom}
\title{Probability calculator for the binomial distribution (binomial)}
\usage{
prob_binom(n, p, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{n}{Number of trials}

\item{p}{Probability}

\item{lb}{Lower bound on the number of successes}

\item{ub}{Upper bound on the number of successes}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the binomial distribution (binomial)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}

