% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_calc}
\alias{prob_calc}
\title{Probability calculator}
\usage{
prob_calc(dist = "normal", mean = 0, stdev = 1, lb = -Inf, ub = Inf)
}
\arguments{
\item{dist}{Distribution}

\item{mean}{Mean}

\item{stdev}{Standard deviation}

\item{lb}{Lower bound}

\item{ub}{Upper bound}
}
\value{
Probability stuff
}
\description{
Probability calculator
}
\details{
See \url{http://vnijs.github.io/radiant/quant/probability.html} for an example in Radiant
}

