% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_disc}
\alias{prob_disc}
\title{Probability calculator for the discrete distribution (discrete)}
\usage{
prob_disc(v, p, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{v}{Values}

\item{p}{Probabilities}

\item{lb}{Lower bound on the number of successes}

\item{ub}{Upper bound on the number of successes}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the discrete distribution (discrete)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}

