% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_pois}
\alias{prob_pois}
\title{Probability calculator for the poisson distribution}
\usage{
prob_pois(lambda, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{lambda}{Rate}

\item{lb}{Lower bound (default is -Inf)}

\item{ub}{Upper bound (default is Inf)}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the poisson distribution
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}

