% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_unif}
\alias{prob_unif}
\title{Probability calculator for the uniform distribution}
\usage{
prob_unif(min, max, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{min}{Minmum value}

\item{max}{Maximum value}

\item{lb}{Lower bound}

\item{ub}{Upper bound}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the uniform distribution
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}

