% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{publishers}
\alias{publishers}
\title{Comic publishers}
\format{A data frame with 3 rows and 2 variables}
\usage{
data(publishers)
}
\description{
Comic publishers
}
\details{
List of comic publishers from \url{http://stat545-ubc.github.io/bit001_dplyr-cheatsheet.html}. The dataset is used to illustrate data merging / joining. Description provided in attr(publishers,"description")
}
\keyword{datasets}

