% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant}
\alias{radiant}
\title{Launch Radiant in the default browser}
\usage{
radiant(app = c("marketing", "quant", "base"))
}
\arguments{
\item{app}{Choose the app to run. Either "base", "quant", or "marketing". "marketing" is the default}
}
\description{
Launch Radiant in the default browser
}
\details{
See \url{http://mostly-harmless.github.io/radiant} for documentation and tutorials
}

