% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant_url}
\alias{radiant_url}
\title{Start radiant but do not open a browser}
\usage{
radiant_url(state, ...)
}
\arguments{
\item{state}{Path to statefile to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Start radiant but do not open a browser
}
\examples{
\dontrun{
radiant_url()
}
}
