% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant_viewer}
\alias{radiant_viewer}
\title{Launch Radiant in the Rstudio Viewer}
\usage{
radiant_viewer()
}
\description{
Launch Radiant in the Rstudio Viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant::radiant_viewer()
}
}
