% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant_window}
\alias{radiant_window}
\title{Launch radiant in an Rstudio window}
\usage{
radiant_window()
}
\description{
Launch radiant in an Rstudio window
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant_window()
}
}
