% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulater.R
\name{repeater}
\alias{repeater}
\title{Repeat simulation}
\usage{
repeater(nr = 12, vars = "", seed = "", sim = "")
}
\arguments{
\item{nr}{Number times to repeat the simulation}

\item{vars}{Variables to use in repeated simulation}

\item{seed}{To repeat a simulation with the same randomly generated values enter a number into Random seed input box.}

\item{sim}{Return value from the simulater function}
}
\description{
Repeat simulation
}

