% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{rndnames}
\alias{rndnames}
\title{100 random names}
\format{A data frame with 100 rows and 2 variables}
\usage{
data(rndnames)
}
\description{
100 random names
}
\details{
A list of 100 random names generated by \url{listofrandomnames.com}.  Description provided in attr(rndnames,"description")
}
\keyword{datasets}

