% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{Simple random sampling}
\usage{
sampling(dataset, smp_var, smp_sample_size, data_filter = "",
  smp_print_full = TRUE)
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{smp_var}{The variable to sample from}

\item{smp_sample_size}{Number of units to select}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}

\item{smp_print_full}{Print full sampling frame. Default is TRUE}
}
\value{
A list of variables defined in sampling as an object of class sampling
}
\description{
Simple random sampling
}
\details{
See \url{http://vnijs.github.io/radiant/quant/sampling.html} for an example in Radiant
}
\examples{
result <- sampling("rndnames","Names",10)
}
\seealso{
\code{\link{summary.sampling}} to summarize results
}

