% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_factor.R
\name{save_factors}
\alias{save_factors}
\title{Deprecated function to store factor loadings}
\usage{
save_factors(object, ..., name = "")
}
\arguments{
\item{object}{Return value from \code{\link{full_factor}}}

\item{...}{Additional arguments}

\item{name}{Name of factor score variables}
}
\description{
Deprecated function to store factor loadings
}
\seealso{
Use \code{\link{store.full_factor}} instead
}

