% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{save_glm_resid}
\alias{save_glm_resid}
\title{Save glm residuals}
\usage{
save_glm_resid(result)
}
\arguments{
\item{result}{Return value from \code{\link{glm_reg}}}
}
\description{
Save glm residuals
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}
\examples{
result <- glm_reg("titanic", "survived", "pclass", glm_levels = "Yes")
save_glm_resid(result)
}

