% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{save_kmeans}
\alias{save_kmeans}
\title{Save cluster means to a csv file}
\usage{
save_kmeans(object, file = "kmeans.csv")
}
\arguments{
\item{object}{Return value from \code{\link{kmeans_clus}}}

\item{file}{Filename and path to use}
}
\description{
Save cluster means to a csv file
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/kmeans_clus.html} for an example in Radiant
}
\examples{
result <- kmeans_clus("shopping", km_vars = c("v1:v6"))
save_kmeans(result, file = "~/shopping_kmeans.csv")
}
\seealso{
\code{\link{kmeans_clus}} to generate results

\code{\link{summary.kmeans_clus}} to summarize results

\code{\link{plot.kmeans_clus}} to plot results

\code{\link{save_membership}} to add cluster membership to the selected dataset
}

