% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{save_membership}
\alias{save_membership}
\title{Deprecated function to store cluster membership}
\usage{
save_membership(object, ..., name = "")
}
\arguments{
\item{object}{Return value from \code{\link{kmeans_clus}}}

\item{...}{Additional arguments}

\item{name}{Name of cluster membership variable}
}
\description{
Deprecated function to store cluster membership
}
\seealso{
Use \code{\link{store.kmeans_clus}} instead
}

