% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{saver}
\alias{saver}
\title{Save data.frame as an rda or rds file from Radiant}
\usage{
saver(objname, file)
}
\arguments{
\item{objname}{Name of the data.frame}

\item{file}{File name and path as a string. Extension must be either rda or rds}
}
\value{
Data.frame in r_data
}
\description{
Save data.frame as an rda or rds file from Radiant
}

