% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore.R
\name{sd_rm}
\alias{sd_rm}
\title{Standard deviation with na.rm = TRUE}
\usage{
sd_rm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Standard deviation
}
\description{
Standard deviation with na.rm = TRUE
}
\examples{
sd_rm(rnorm(100))
}

