% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{sdw}
\alias{sdw}
\title{Standard deviation of weighted sum of variables}
\usage{
sdw(...)
}
\arguments{
\item{...}{A matched number of weights and stocks}
}
\value{
A vector of standard deviation estimates
}
\description{
Standard deviation of weighted sum of variables
}

