% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{serr}
\alias{serr}
\title{Standard error}
\usage{
serr(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Standard error
}
\description{
Standard error
}
\examples{
serr(rnorm(100))

}

