% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{set_class}
\alias{set_class}
\title{Alias to set class for an object (i.e., the main analysis functions)}
\usage{
set_class()
}
\description{
Alias to set class for an object (i.e., the main analysis functions)
}
\examples{
foo <- function(x) x^2 \%>\% set_class(c("foo", class(.)))
}

