% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{shopping}
\alias{shopping}
\title{Shopping attitudes}
\format{A data frame with 20 rows and 7 variables}
\usage{
data(shopping)
}
\description{
Shopping attitudes
}
\details{
Attitudinal data on shopping for 20 consumers. Description provided in attr(shopping,"description")
}
\keyword{datasets}

