% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{sig_stars}
\alias{sig_stars}
\title{Add "***" to a data.frame based on the p.values}
\usage{
sig_stars(p.value)
}
\arguments{
\item{p.value}{Vector of p.values from an analysis}
}
\value{
A list with all variables defined in the function
}
\description{
Add "***" to a data.frame based on the p.values
}
\examples{
sig_stars(c(.0009,.009, .049, .09, .4))
}

