% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{sim_splitter}
\alias{sim_splitter}
\title{Split input command string}
\usage{
sim_splitter(x, symbol = " ")
}
\arguments{
\item{x}{Input string}

\item{symbol}{Symbol used to split the command string}
}
\value{
Split input command string
}
\description{
Split input command string
}

