% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R, R/simulater.R
\name{simulater}
\alias{simulater}
\title{Simulate data for decision analysis}
\usage{
simulater(const = "", norm = "", unif = "", discrete = "", form = "",
  nr = 1000, seed = "1234", name = "sim1")

simulater(const = "", norm = "", unif = "", discrete = "", form = "",
  nr = 1000, seed = "1234", name = "sim1")
}
\arguments{
\item{const}{A string listing the constants to include in the analysis (e.g., "cost = 3\\nsize = 4")}

\item{norm}{A string listing the normally distributed random variables to include in the analysis (e.g., "demand = c(2000,1000)" where the first number is the mean and the second is the standard deviation)}

\item{unif}{A string listing the uniformly distributed random variables to include in the analysis (e.g., "demand = c(0,1)" where the first number is the minimum value and the second is the maximum value)}

\item{discrete}{A string listing the random variables with a discrete distribution to include in the analysis (e.g., "price = c(5, .3, 8, .7)" where for each pair of numbers the first is the value and the second the probability}

\item{form}{A string for the formula to evaluate (e.g., "profit = demand * (price - cost)")}

\item{nr}{Number of simulation runs}

\item{seed}{Random seed to use for simulation}

\item{name}{Name of the dataset to create when run inside Radiant}

\item{const}{A string listing the constants to include in the analysis (e.g., "cost = 3\\n = 4")}

\item{norm}{A string listing the normally distributed random variables to include in the analysis (e.g., "demand = c(2000,1000)" where the first number is the mean and the second is the standard deviation)}

\item{unif}{A string listing the uniformly distributed random variables to include in the analysis (e.g., "demand = c(0,1)" where the first number is the minimum value and the second is the maximum value)}

\item{discrete}{A string listing the random variables with a discrete distribution to include in the analysis (e.g., "price = c(5, .3, 8, .7)" where for each pair of numbers the first is the value and the second the probability}

\item{form}{A string for the formula to evaluate (e.g., "profit = demand * (price - cost)")}
}
\value{
A data.frame with the created variables

A data.frame with the created variables
}
\description{
Simulate data for decision analysis

Simulate data for decision analysis
}
\details{
See \url{http://vnijs.github.io/radiant/base/simulater.html} for an example in Radiant

See \url{http://vnijs.github.io/radiant/base/simulater.html} for an example in Radiant
}
\examples{
result <- simulater(const = "cost = 3", norm = "demand = c(2000,1000)",
                    discrete = "price = c(5, .3, 8, .7)",
                    form = "profit = demand * (price - cost)")
result <- simulater(const = "cost = 3", norm = "demand = c(2000,1000)",
                    discrete = "price = c(5, .3, 8, .7)",
                    form = "profit = demand * (price - cost)")
}
\seealso{
\code{\link{summary.simulater}} to summarize results

\code{\link{plot.simulater}} to plot results

\code{\link{summary.simulater}} to summarize results

\code{\link{plot.simulater}} to plot results
}

