% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_mean.R
\name{single_mean}
\alias{single_mean}
\title{Compare a sample mean to a population mean}
\usage{
single_mean(dataset, sm_var, sm_comp_value = 0,
  sm_alternative = "two.sided", sm_sig_level = 0.95)
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in a values list from Radiant}

\item{sm_var}{The variable selected for the mean comparison}

\item{sm_comp_value}{Population value to compare the sample mean with}

\item{sm_alternative}{The alternative hypothesis (two.sided, greater or less)}

\item{sm_sig_level}{Span of the confidence interval}
}
\value{
A list with all variables defined in the function
}
\description{
Compare a sample mean to a population mean
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_mean.html} for an example in Radiant
}
\examples{
single_mean("diamonds","price")
}
\seealso{
\code{\link{summary_single_mean}} to print results

\code{\link{plots_single_mean}} to plot results
}

