% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_prop.R
\name{single_prop}
\alias{single_prop}
\title{Compare a sample proportion to a population proportion}
\usage{
single_prop(dataset, sp_var, sp_levels, sp_comp_value = 0.5,
  sp_alternative = "two.sided", sp_sig_level = 0.95, sp_plots = "hist")
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{sp_var}{The variable selected for the proportion comparison}

\item{sp_comp_value}{Population value to compare the sample proportion to}

\item{sp_alternative}{The alternative hypothesis (two.sided, greater or less)}

\item{sp_sig_level}{Span of the confidence interval}
}
\value{
A list with all variables defined in the function as an object of class single_prop
}
\description{
Compare a sample proportion to a population proportion
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_prop.html} for an example in Radiant
}
\examples{
single_prop("diamond","cut","IF")
}
\seealso{
\code{\link{summary.single_prop}} to summarize results

\code{\link{plots.single_prop}} to plot results
}

