% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{standardize}
\alias{standardize}
\title{Standardize}
\usage{
standardize(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
If x is a numberic variable return center(x) / mean(x)
}
\description{
Standardize
}

