% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{state_init}
\alias{state_init}
\title{Set initial value for shiny input}
\usage{
state_init(inputvar, init = "")
}
\arguments{
\item{inputvar}{Name shiny input}

\item{init}{Initial value to use if state value for input not set}
}
\value{
value for inputvar
}
\description{
Set initial value for shiny input
}
\details{
Useful for radio button or checkbox
}
\examples{
r_state <<- list()
state_init("test")
state_init("test",0)
r_state$test <- c("a","b")
state_init("test",0)
shiny::radioButtons("rb", label = "Button:", c("a","b"), selected = state_init("rb", "a"))
r_state$rb <- "b"
shiny::radioButtons("rb", label = "Button:", c("a","b"), selected = state_init("rb", "a"))
rm(r_state)
}
\seealso{
\code{\link{state_single}}

\code{\link{state_multiple}}

\code{\link{copy_from}}
}

