% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{state_single}
\alias{state_single}
\title{Set initial value for shiny input from a list of values}
\usage{
state_single(inputvar, vals, init = character(0))
}
\arguments{
\item{inputvar}{Name shiny input}

\item{vals}{Possible values for inputvar}

\item{init}{Initial value to use if state value for input not set}
}
\value{
value for inputvar
}
\description{
Set initial value for shiny input from a list of values
}
\details{
Useful for select input with multiple = FALSE
}
\examples{
r_state <- list()
state_single("test",1:10,1)
r_state$test <- 8
state_single("test",1:10,1)
shiny::selectInput("si", label = "Select:", c("a","b"), selected = state_single("si"))
r_state$si <- "b"
shiny::selectInput("si", label = "Select:", c("a","b"), selected = state_single("si", "b"))
}
\seealso{
\code{\link{state_init}}

\code{\link{state_multiple}}

\code{\link{copy_from}}
}

