% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.R
\name{store.ann}
\alias{store.ann}
\title{Store predicted values generated in the ann function}
\usage{
\method{store}{ann}(object, ..., data = "", name = "predict_ann")
}
\arguments{
\item{object}{Return value from predict.nnet}

\item{...}{Additional arguments}

\item{data}{Dataset name}

\item{name}{Variable name assigned to the predicted values}
}
\description{
Store predicted values generated in the ann function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/ann.html} for an example in Radiant
}

