% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_factor.R
\name{store.full_factor}
\alias{store.full_factor}
\title{Store factor scores to active dataset}
\usage{
\method{store}{full_factor}(object, ..., name = "")
}
\arguments{
\item{object}{Return value from \code{\link{full_factor}}}

\item{...}{Additional arguments}

\item{name}{Name of factor score variables}
}
\description{
Store factor scores to active dataset
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/full_factor.html} for an example in Radiant
}
\examples{
\dontrun{
result <- full_factor("diamonds",c("price","carat","table"))
store(result)
head(diamonds)
}

}
\seealso{
\code{\link{full_factor}} to generate results

\code{\link{summary.full_factor}} to summarize results

\code{\link{plot.full_factor}} to plot results
}

