% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{store.glm_predict}
\alias{store.glm_predict}
\title{Store predicted values generated in the glm_reg function}
\usage{
\method{store}{glm_predict}(object, ..., name = "pred_glm")
}
\arguments{
\item{object}{Return value from \code{\link{glm_reg}} or \code{\link{predict.glm_reg}}}

\item{...}{Additional arguments. Must include data or dataset name (e.g., data = mtcars or data = "mtcars")}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Store predicted values generated in the glm_reg function
}
\details{
Use \code{\link{store.glm_predict}} or \code{\link{store.glm_reg}} instead

See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}

