% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{store.glm_reg}
\alias{store.glm_reg}
\title{Store residuals from a model generated in the glm_reg function}
\usage{
\method{store}{glm_reg}(object, ..., name = "residuals_glm")
}
\arguments{
\item{object}{Return value from \code{\link{glm_reg}}}

\item{...}{Additional arguments}

\item{name}{Variable name(s) assigned to predicted values}
}
\description{
Store residuals from a model generated in the glm_reg function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/glm_reg} for an example in Radiant
}

