% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{store.kmeans_clus}
\alias{store.kmeans_clus}
\title{Add a cluster membership variable to the active dataset}
\usage{
\method{store}{kmeans_clus}(object)
}
\arguments{
\item{object}{Return value from \code{\link{kmeans_clus}}}
}
\description{
Add a cluster membership variable to the active dataset
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/kmeans_clus.html} for an example in Radiant
}
\examples{
\dontrun{
result <- kmeans_clus("shopping", vars = c("v1:v6"))
store_kmeans(result)
head(shopping)
}
}
\seealso{
\code{\link{kmeans_clus}} to generate results

\code{\link{summary.kmeans_clus}} to summarize results

\code{\link{plot.kmeans_clus}} to plot results
}

