% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap.R
\name{store.pmap}
\alias{store.pmap}
\title{Store factor scores from attribute based perceptual map}
\usage{
\method{store}{pmap}(object, ..., name = "")
}
\arguments{
\item{object}{Return value from \code{\link{pmap}}}

\item{...}{Additional arguments}

\item{name}{Name of factor score variables}
}
\description{
Store factor scores from attribute based perceptual map
}
\seealso{
Use \code{\link{store.full_factor}} instead
}

