% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{store}
\alias{store}
\title{Method to store variables in a dataset in Radiant}
\usage{
store(object, ...)
}
\arguments{
\item{object}{Object of relevant class that has required information to store}

\item{...}{Additional arguments}
}
\description{
Method to store variables in a dataset in Radiant
}

