% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{store.regression}
\alias{store.regression}
\title{Store residuals from a model generated in the regression function}
\usage{
\method{store}{regression}(object, ..., name = "residuals_reg")
}
\arguments{
\item{object}{Return value from \code{\link{regression}}}

\item{...}{Additional arguments}

\item{name}{Variable name(s) assigned to predicted values}
}
\description{
Store residuals from a model generated in the regression function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}

