% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.R
\name{store_ann}
\alias{store_ann}
\title{Deprecated function to store predictions from an ANN}
\usage{
store_ann(object, ..., data = "", name = "predict_ann")
}
\arguments{
\item{object}{Return value from predict.nnet}

\item{...}{Additional arguments}

\item{data}{Dataset name}

\item{name}{Variable name assigned to the predicted values}
}
\description{
Deprecated function to store predictions from an ANN
}

