% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{store_crs}
\alias{store_crs}
\title{Store predicted values generated in the crs function}
\usage{
store_crs(pred, data, name = "pred_crs")
}
\arguments{
\item{pred}{Return value from predict.nnet}

\item{data}{Dataset name}

\item{name}{Variable name assigned to the predicted values}
}
\description{
Store predicted values generated in the crs function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/crs.html} for an example in Radiant
}

