% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{store_glm}
\alias{store_glm}
\title{Deprecated function to store logistic regression residuals and predictions}
\usage{
store_glm(object, data = object$dataset, type = "residuals",
  name = paste0(type, "_glm"))
}
\arguments{
\item{object}{Return value from \code{\link{glm_reg}} or \code{\link{predict.glm_reg}}}

\item{data}{Dataset name}

\item{type}{Residuals ("residuals") or predictions ("predictions"). For predictions the dataset name must be provided}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Deprecated function to store logistic regression residuals and predictions
}
\details{
Use \code{\link{store.glm_predict}} or \code{\link{store.glm_reg}} instead

See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}

