% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{store_glm_resid}
\alias{store_glm_resid}
\title{Store residuals generated in the glm_reg function}
\usage{
store_glm_resid(object)
}
\arguments{
\item{object}{Return value from \code{\link{glm_reg}}}
}
\description{
Store residuals generated in the glm_reg function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}
\examples{
\donttest{
result <- glm_reg("titanic", "survived", "pclass", lev = "Yes")
store_glm_resid(result)
head(titanic)
}
}

