% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{sum_rm}
\alias{sum_rm}
\title{Sum with na.rm = TRUE}
\usage{
sum_rm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Sum of input values
}
\description{
Sum with na.rm = TRUE
}
\examples{
sum_rm(1:200)

}

