% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ann.R
\name{summary.ann}
\alias{summary.ann}
\title{Summary method for the ann function}
\usage{
\method{summary}{ann}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{ann}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the ann function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/ann.html} for an example in Radiant
}
\examples{
result <- ann("titanic", "survived", "pclass", lev = "Yes")
summary(result)
}
\seealso{
\code{\link{ann}} to generate the results

\code{\link{plot.ann}} to plot the results
}

