% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means.R
\name{summary.compare_means}
\alias{summary.compare_means}
\title{Summary method for the compare_means function}
\usage{
\method{summary}{compare_means}(object, show = FALSE, ...)
}
\arguments{
\item{object}{Return value from \code{\link{compare_means}}}

\item{show}{Show additional output (i.e., t.value, df, and confidence interval)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the compare_means function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
result <- compare_means("diamonds","cut","price")
summary(result)
result <- diamonds \%>\% tbl_df \%>\% compare_means("x","y")
summary(result)
result <- diamonds \%>\% tbl_df \%>\% group_by(cut) \%>\% compare_means("x",c("x","y"))
summary(result)

}
\seealso{
\code{\link{compare_means}} to calculate results

\code{\link{plot.compare_means}} to plot results
}

