% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint.R
\name{summary.conjoint}
\alias{summary.conjoint}
\title{Summary method for conjoint output}
\usage{
\method{summary}{conjoint}(result, ca_vif = FALSE)
}
\arguments{
\item{result}{Return value from \code{\link{conjoint}}}

\item{ca_vif}{Shows multicollinearity diagnostics.}
}
\description{
Summary method for conjoint output
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/conjoint.html} for an example in Radiant
}
\examples{
result <- conjoint(dataset = "mp3", ca_dep_var = "Rating", ca_indep_var = "Memory:Shape")
summary(result, ca_vif = TRUE)
}
\seealso{
\code{\link{conjoint}} to generate results

\code{\link{plot.conjoint}} to plot results
}

