% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation.R
\name{summary.correlation}
\alias{summary.correlation}
\title{Summarize method for output from the correlation function. This is a method of class correlation and can be called as summary or summary.correlation}
\usage{
\method{summary}{correlation}(result)
}
\description{
Summarize method for output from the correlation function. This is a method of class correlation and can be called as summary or summary.correlation
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/correlation.html} for an example in Radiant
}
\examples{
result <- correlation("titanic", "pclass", "survived")
summary(result)
}
\seealso{
\code{\link{correlation}} to calculate results

\code{\link{plot.correlation}} to plot results
}

