% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cross_tabs.R
\name{summary.cross_tabs}
\alias{summary.cross_tabs}
\title{Summarize method for output from the cross_tabs function. This is a method of class cross_tabs and can be called as summary or summary.cross_tabs}
\usage{
\method{summary}{cross_tabs}(result)
}
\description{
Summarize method for output from the cross_tabs function. This is a method of class cross_tabs and can be called as summary or summary.cross_tabs
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/cross_tabs.html} for an example in Radiant
}
\examples{
result <- cross_tabs("titanic", "pclass", "survived")
summary(result)
}
\seealso{
\code{\link{cross_tabs}} to calculate results

\code{\link{plot.cross_tabs}} to plot results
}

