% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe.R
\name{summary.doe}
\alias{summary.doe}
\title{Summary method for doe function}
\usage{
\method{summary}{doe}(object, eff = TRUE, part = TRUE, full = TRUE, ...)
}
\arguments{
\item{object}{Return value from \code{\link{conjoint_profiles}}}

\item{eff}{If TRUE print efficiency output}

\item{part}{If TRUE print partial factorial}

\item{full}{If TRUE print full factorial}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary method for doe function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/doe.html} for an example in Radiant
}
\examples{
"price; $10; $13; $16\\nfood; popcorn; gourmet; no food" \%>\% doe \%>\% summary

}
\seealso{
\code{\link{doe}} to calculate results
}

