% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtree.R
\name{summary.dtree}
\alias{summary.dtree}
\title{Summary method for the dree function}
\usage{
\method{summary}{dtree}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{simulater}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the dree function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/dtree.html} for an example in Radiant
}
\seealso{
\code{\link{dtree}} to generate the results

\code{\link{plot.dtree}} to plot results
}

