% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{summary.glm_reg}
\alias{summary.glm_reg}
\title{Summary method for glm_reg}
\usage{
\method{summary}{glm_reg}(result, savepred = FALSE)
}
\arguments{
\item{result}{Return value from \code{\link{glm_reg}}}
}
\description{
Summary method for glm_reg
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/glm.html} for an example in Radiant
}
\examples{
result <- glm_reg("titanic", "survived", c("pclass","sex"), glm_levels = "Yes")
summary(result)
}
\seealso{
\code{\link{glm_reg}} to generate the results

\code{\link{plot.glm_reg}} to plot the results
}

