% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clus.R
\name{summary.hier_clus}
\alias{summary.hier_clus}
\title{Summary method for the hier_clus function}
\usage{
\method{summary}{hier_clus}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{hier_clus}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the hier_clus function
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/hier_clus.html} for an example in Radiant
}
\examples{
result <- hier_clus("shopping", vars = c("v1:v6"))
summary(result)

}
\seealso{
\code{\link{hier_clus}} to generate results

\code{\link{plot.hier_clus}} to plot results
}

