% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{summary.kmeans_clus}
\alias{summary.kmeans_clus}
\title{Summary method for kmeans_clus}
\usage{
\method{summary}{kmeans_clus}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{kmeans_clus}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for kmeans_clus
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/kmeans_clus.html} for an example in Radiant
}
\examples{
result <- kmeans_clus("shopping", km_vars = c("v1:v6"))
summary(result)
shopping \%>\% kmeans_clus(km_vars = c("v1:v6"), km_nr_clus = 3) \%>\% summary
}
\seealso{
\code{\link{kmeans_clus}} to generate results

\code{\link{plot.kmeans_clus}} to plot results

\code{\link{save_membership}} to add cluster membership to the selected dataset
}

