% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mds.R
\name{summary.mds}
\alias{summary.mds}
\title{Summary method for the mds function}
\usage{
\method{summary}{mds}(object, mds_round = 1, ...)
}
\arguments{
\item{object}{Return value from \code{\link{mds}}}

\item{mds_round}{Rounding to use for output (default = 0). +1 used for coordinates. +2 used for stress measure. Not currently accessible in Radiant}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the mds function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/mds.html} for an example in Radiant
}
\examples{
result <- mds("city","from","to","distance")
summary(result)
summary(result, mds_round = 2)
}
\seealso{
\code{\link{mds}} to calculate results

\code{\link{plot.mds}} to plot results
}

