% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{summary.performance}
\alias{summary.performance}
\title{Summary method for the performance function}
\usage{
\method{summary}{performance}(object, prn = TRUE, ...)
}
\arguments{
\item{object}{Return value from \code{\link{performance}}}

\item{prn}{Print model performance results (default is TRUE)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the performance function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/performance.html} for an example in Radiant
}
\examples{
performance("titanic", "age", "survived") \%>\% summary
performance("titanic", c("age","fare"), "survived") \%>\% summary
}
\seealso{
\code{\link{performance}} to summarize results

\code{\link{plot.performance}} to plot results
}

