% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pmap.R
\name{summary.pmap}
\alias{summary.pmap}
\title{Summary method for the pmap function}
\usage{
\method{summary}{pmap}(object, cutoff = 0, ...)
}
\arguments{
\item{object}{Return value from \code{\link{pmap}}}

\item{cutoff}{Show only loadings with (absolute) values above cutoff (default = 0)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the pmap function
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/pmap.html} for an example in Radiant
}
\examples{
result <- pmap("computer","brand","high_end:business")
summary(result)
summary(result, cutoff = .3)
result <- pmap("computer","brand","high_end:dated", pref = c("innovative","business"))
summary(result)
computer \%>\% pmap("brand","high_end:dated", pref = c("innovative","business")) \%>\%
  summary
}
\seealso{
\code{\link{pmap}} to calculate results

\code{\link{plot.pmap}} to plot results
}

