% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_disc}
\alias{summary.prob_disc}
\title{Summary method for the probability calculator function (discrete)}
\usage{
\method{summary}{prob_disc}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_disc}}}

\item{type}{Probabilities or values}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator function (discrete)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}

