% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_expo}
\alias{summary.prob_expo}
\title{Summary method for the probability calculator function (Exponential distribution)}
\usage{
\method{summary}{prob_expo}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_expo}}}

\item{type}{Probabilities or values}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator function (Exponential distribution)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}

