% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_unif}
\alias{summary.prob_unif}
\title{Summary method for the probability calculator function (normal)}
\usage{
\method{summary}{prob_unif}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_unif}}}

\item{type}{Probabilities or values}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator function (normal)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/probability.html} for an example in Radiant
}

