% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression.R
\name{summary.regression}
\alias{summary.regression}
\title{Summarize results from the regression. This is a method of class regression and can be called as summary or summary.regression}
\usage{
\method{summary}{regression}(result)
}
\arguments{
\item{result}{Return value from \code{\link{regression}}}
}
\description{
Summarize results from the regression. This is a method of class regression and can be called as summary or summary.regression
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
result <- regression("diamonds", "price", c("carat","clarity"))
summary(result)
}
\seealso{
\code{\link{regression}} to generate the results

\code{\link{plot.regression}} to plot results
}

