% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulater.R
\name{summary.repeater}
\alias{summary.repeater}
\title{Summarize repeated simulation}
\usage{
\method{summary}{repeater}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{simulater}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summarize repeated simulation
}

