% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_comp.R
\name{summary.sample_size_comp}
\alias{summary.sample_size_comp}
\title{Summary method for the sample_size_comp function}
\usage{
\method{summary}{sample_size_comp}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{sample_size_comp}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the sample_size_comp function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/sample_size_comp} for an example in Radiant
}
\seealso{
\code{\link{sample_size_comp}} to generate the results
}

