% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampling.R
\name{summary.sampling}
\alias{summary.sampling}
\title{Summary method for sampling}
\usage{
\method{summary}{sampling}(result)
}
\arguments{
\item{result}{Return value from \code{\link{sampling}}}
}
\description{
Summary method for sampling
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/sampling} for an example in Radiant
}
\examples{
result <- sampling("rndnames","Names",10)
summary(result)
}
\seealso{
\code{\link{sampling}} to generate the results
}

