% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulater.R
\name{summary.simulater}
\alias{summary.simulater}
\title{Summary method for the simulater function}
\usage{
\method{summary}{simulater}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{simulater}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the simulater function
}
\details{
See \url{http://vnijs.github.io/radiant/base/simulater.html} for an example in Radiant
}
\examples{
result <- simulater(norm = "demand 2000 1000")
summary(result)
}
\seealso{
\code{\link{simulater}} to generate the results

\code{\link{plot.simulater}} to plot results
}

