% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_prop.R
\name{summary.single_prop}
\alias{summary.single_prop}
\title{Summarize results from the single_prop function. This is a method of class single_prop and can be called as summary or summary.single_prop}
\usage{
\method{summary}{single_prop}(result)
}
\arguments{
\item{result}{Return value from \code{\link{single_prop}}}
}
\description{
Summarize results from the single_prop function. This is a method of class single_prop and can be called as summary or summary.single_prop
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_prop.html} for an example in Radiant
}
\examples{
result <- single_prop("diamond","cut")
summary(result)
}
\seealso{
\code{\link{single_prop}} to generate the results

\code{\link{plots.single_prop}} to plot results
}

