% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_mean.R
\name{summary_single_mean}
\alias{summary_single_mean}
\title{Summarize results from the single_mean function}
\usage{
summary_single_mean(result)
}
\arguments{
\item{result}{Return value from \code{\link{single_mean}}}
}
\description{
Summarize results from the single_mean function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_mean.html} for an example in Radiant
}
\examples{
result <- single_mean("diamonds","price")
summary_single_mean(result)
}
\seealso{
\code{\link{single_mean}} to generate the result

\code{\link{plots_single_mean}} to plot results
}

