% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{superheroes}
\alias{superheroes}
\title{Super heroes}
\format{A data frame with 7 rows and 4 variables}
\usage{
data(superheroes)
}
\description{
Super heroes
}
\details{
List of super heroes from \url{http://stat545-ubc.github.io/bit001_dplyr-cheatsheet.html}. The dataset is used to illustrate data merging / joining. Description provided in attr(superheroes,"description")
}
\keyword{datasets}

