% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression.R
\name{test_check}
\alias{test_check}
\title{Add interaction terms to list of test variables if needed}
\usage{
test_check(test_var, int_var)
}
\arguments{
\item{test_var}{List of variables to use for testing for _regression_ or _glm_}

\item{int_var}{Interaction terms specified}
}
\value{
'test_var' is a vector of variables to test
}
\description{
Add interaction terms to list of test variables if needed
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
test_check("a", c("a:b","b:c"))
}

