% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{test_specs}
\alias{test_specs}
\title{Add interaction terms to list of test variables if needed}
\usage{
test_specs(test_var, int)
}
\arguments{
\item{test_var}{List of variables to use for testing for regression or glm_reg}

\item{int}{Interaction terms specified}
}
\value{
A vector of variables names to test
}
\description{
Add interaction terms to list of test variables if needed
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
test_specs("a", c("a:b", "b:c"))

}

