% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Survival data for the Titanic}
\format{A data frame with 1309 rows and 11 variables}
\usage{
data(titanic)
}
\description{
Survival data for the Titanic
}
\details{
Survival data for the Titanic. Description provided in attr(titanic,"description")
}
\keyword{datasets}

