% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{titanic_pred}
\alias{titanic_pred}
\title{Predict survival}
\format{A data frame with 6 rows and 3 variables}
\usage{
data(titanic_pred)
}
\description{
Predict survival
}
\details{
Prediction data.frame for glm_reg based on the Titanic dataset
}
\keyword{datasets}

