% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{toothpaste}
\alias{toothpaste}
\title{Toothpaste attitudes}
\format{A data frame with 60 rows and 10 variables}
\usage{
data(toothpaste)
}
\description{
Toothpaste attitudes
}
\details{
Attitudinal data on toothpaste for 60 consumers. Description provided in attr(toothpaste,"description")
}
\keyword{datasets}

