% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{varp_rm}
\alias{varp_rm}
\title{Variance for the population na.rm = TRUE}
\usage{
varp_rm(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Variance for the population
}
\description{
Variance for the population na.rm = TRUE
}
\examples{
varp_rm(rnorm(100))

}

