% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{weighted.sd}
\alias{weighted.sd}
\title{Weighted standard deviation}
\usage{
weighted.sd(x, wt, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{wt}{Numeric vector of weights}

\item{na.rm}{Remove missing values (default is TRUE)}
}
\description{
Weighted standard deviation
}
\details{
Calculated a weighted standard deviation
}

