% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{which.pmax}
\alias{which.pmax}
\title{Returns the index of the (parallel) maxima of the input values}
\usage{
which.pmax(...)
}
\arguments{
\item{...}{Numeric or character vectors of the same length}
}
\value{
Vector of rankings
}
\description{
Returns the index of the (parallel) maxima of the input values
}
\examples{
which.pmax(1:10, 10:1)
which.pmax(2, 10:1)

}

