% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{which.pmin}
\alias{which.pmin}
\title{Returns the index of the (parallel) minima of the input values}
\usage{
which.pmin(...)
}
\arguments{
\item{...}{Numeric or character vectors of the same length}
}
\value{
Vector of rankings
}
\description{
Returns the index of the (parallel) minima of the input values
}
\examples{
which.pmin(1:10, 10:1)
which.pmin(2, 10:1)

}

