% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{win_launcher}
\alias{win_launcher}
\title{Create a .bat launcher for Windows}
\usage{
win_launcher(app = c("marketing", "quant", "base"))
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("marketing", "quant", or "base"). Default is "marketing"}
}
\value{
On Windows a file `radiant.bat` will be put on the desktop. Double-click the file to launch the specified Radiant app
}
\description{
Create a .bat launcher for Windows
}

