% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{xtile}
\alias{xtile}
\title{Create a quintile (or decile) index}
\usage{
xtile(x, n)
}
\arguments{
\item{x}{Numeric variable}

\item{n}{number of bins to create}

\item{ties}{How to break ties ("first","last","random")}
}
\description{
Create a quintile (or decile) index
}
\details{
Same as stata
}
\examples{
xtile(1:10,5)
}

