% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{changedata}
\alias{changedata}
\title{Change data}
\usage{
changedata(dataset, vars = c(), var_names = names(vars))
}
\arguments{
\item{dataset}{Name of the dataframe to change}

\item{vars}{New variables to add to the data.frame}

\item{var_names}{Names for the new variables to add to the data.frame}
}
\value{
None
}
\description{
Change data
}

