% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{ci_perc}
\alias{ci_perc}
\title{Values at confidence levels}
\usage{
ci_perc(dat, alt = "two.sided", cl = 0.95)
}
\arguments{
\item{dat}{Data}

\item{alt}{Type of hypothesis ("two.sided","less","greater")}

\item{cl}{Confidence level}
}
\value{
A vector with values at a confidence level
}
\description{
Values at confidence levels
}
\examples{
ci_perc(0:100, "less",.95)
ci_perc(0:100, "greater",.95)
ci_perc(0:100, "two.sided",.80)

}

