% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{city}
\alias{city}
\title{Distance in miles between nine cities in the USA
The dataset is used to illustrate multi-dimensional scaling (MDS)
Description provided in attr(city,"description")}
\format{A data frame with 45 rows and 3 variables}
\usage{
data(city)

data(city)
}
\description{
Distance in miles between nine cities in the USA
The dataset is used to illustrate multi-dimensional scaling (MDS)
Description provided in attr(city,"description")

Distance in miles between nine cities in the USA
The dataset is used to illustrate multi-dimensional scaling (MDS)
Description provided in attr(city,"description")
}
\keyword{datasets}

