% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/full_factor.R
\name{clean_loadings}
\alias{clean_loadings}
\title{Sort and clean loadings}
\usage{
clean_loadings(ff_loadings, ff_cutoff = 0, ff_sort = FALSE, ff_round = 8)
}
\arguments{
\item{ff_loadings}{Data.frame with loadings}

\item{ff_cutoff}{Show only loadings with (absolute) values above ff_cutoff (default = 0)}

\item{ff_sort}{Sort factor loadings}

\item{ff_round}{Number of digits to show}
}
\description{
Sort and clean loadings
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/full_factor.html} for an example in Radiant
}
\examples{
result <- full_factor("diamonds",c("price","carat","table","x","y"))
clean_loadings(result$ff_loadings, TRUE, .5, 2)
}

