% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means.R
\name{compare_means}
\alias{compare_means}
\title{Compare means for two or more variables}
\usage{
compare_means(dataset, var1, var2, samples = "independent",
  alternative = "two.sided", conf_lev = 0.95, comb = "",
  adjust = "none", test = "t", dec = 3, data_filter = "")
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{var1}{A numeric variable or factor selected for comparison}

\item{var2}{One or more numeric variables for comparison. If var1 is a factor only one variable can be selected and the mean of this variable is compared across (factor) levels of va1r}

\item{samples}{Are samples independent ("independent") or not ("paired")}

\item{alternative}{The alternative hypothesis ("two.sided", "greater" or "less")}

\item{conf_lev}{Span of the confidence interval}

\item{comb}{Combinations to evaluate}

\item{adjust}{Adjustment for multiple comparisons ("none" or "bonf" for Bonferroni)}

\item{test}{t-test ("t") or Wilcox ("wilcox")}

\item{dec}{Number of decimals to show}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}
}
\value{
A list of all variables defined in the function as an object of class compare_means
}
\description{
Compare means for two or more variables
}
\details{
See \url{http://vnijs.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
result <- compare_means("diamonds","cut","price")
result <- diamonds \%>\% compare_means("cut","price")

}
\seealso{
\code{\link{summary.compare_means}} to summarize results

\code{\link{plot.compare_means}} to plot results
}

