% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint_profiles.R
\name{conjoint_profiles}
\alias{conjoint_profiles}
\title{Create fractional factorial design for conjoint analysis}
\usage{
conjoint_profiles(dataset)
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}
}
\value{
A list with all variables defined in the function as an object of class conjoint_profiles
}
\description{
Create fractional factorial design for conjoint analysis
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/conjoint_profiles.html} for an example in Radiant
}
\examples{
ca_prof <- readLines(system.file("examples/profiles-movie.txt", package='radiant'))
result <- conjoint_profiles("ca_prof")
}
\seealso{
\code{\link{summary.conjoint_profiles}} to summarize results
}

