% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore data}
\usage{
explore(dataset, expl_vars = "", data_filter = "", expl_byvar = "",
  expl_fun = c("length", "mean_rm"))
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{expl_vars}{(Numerical) variables to summaries}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}

\item{expl_byvar}{Variable(s) to group data by before summarizing}

\item{expl_fun}{Functions to use for summarizing}
}
\value{
A list of all variables defined in the function as an object of class explore
}
\description{
Explore data
}
\details{
See \url{http://vnijs.github.io/radiant/base/explore.html} for an example in Radiant
}
\examples{
result <- explore("diamonds", "price:x")
summary(result)
result <- explore("diamonds", "price", expl_byvar = "cut", expl_fun = c("length", "skew"))
summary(result)
}
\seealso{
\code{\link{summary.explore}} to show summaries

\code{\link{plot.explore}} to plot summaries
}

