% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{getdata}
\alias{getdata}
\title{Get data for analysis functions}
\usage{
getdata(dataset, vars = "", filt = "", rows = NULL, na.rm = TRUE)
}
\arguments{
\item{dataset}{Name of the dataframe}

\item{vars}{Variables to extract from the dataframe}

\item{filt}{Filter to apply to the specified dataset. For example "price > 10000" if dataset is "diamonds" (default is "")}

\item{rows}{Select rows in the specified dataset. For example "1:10" for the first 10 rows or "n()-10:n()" for the last 10 rows (default is NULL)}

\item{na.rm}{Remove rows with missing values (default is TRUE)}
}
\value{
Data.frame with specified columns and rows
}
\description{
Get data for analysis functions
}

