% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{getdata_exp}
\alias{getdata_exp}
\title{Get data for analysis functions exported by Radiant}
\usage{
getdata_exp(dataset, vars, na.rm = TRUE)
}
\arguments{
\item{dataset}{Name of the data.frame}

\item{vars}{Variables to extract from the data.frame}

\item{na.rm}{Remove rows with missing values (default is TRUE)}
}
\value{
Data.frame with the specified columns selected
}
\description{
Get data for analysis functions exported by Radiant
}

