% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{getsummary}
\alias{getsummary}
\title{Create data.frame summary}
\usage{
getsummary(dat, dc = getclass(dat))
}
\arguments{
\item{dat}{Data.frame}

\item{dc}{Class for each variable}
}
\description{
Create data.frame summary
}
\details{
Used in Radiant's Data > Transform tab
}

