% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{lin_launcher}
\alias{lin_launcher}
\title{Create a launcher and updater for Linux (.sh)}
\usage{
lin_launcher(app = c("radiant", "radiant.data", "radiant.design",
  "radiant.basics", "radiant.model", "radiant.multivariate"))
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("analytics", "marketing", "quant", or "base"). Default is "analytics"}
}
\description{
Create a launcher and updater for Linux (.sh)
}
\details{
On Linux a file named 'radiant.sh' and one named 'update_radiant.sh' will be put on the desktop. Double-click the file to launch the specified Radiant app or update Radiant to the latest version
}

