% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{loadcsv}
\alias{loadcsv}
\title{Load a csv file with read.csv and read_csv}
\usage{
loadcsv(fn, .csv = FALSE, header = TRUE, sep = ",", dec = ".",
  saf = TRUE, safx = 20)
}
\arguments{
\item{fn}{File name string}

\item{.csv}{Use read.csv instead of read_csv to load file (default is FALSE)}

\item{header}{Header in file (TRUE, FALSE)}

\item{sep}{Use , (default) or ; or \\t}

\item{dec}{Decimal symbol. Use . (default) or ,}

\item{saf}{Convert character variables to factors if (1) there are less than 100 distinct values (2) there are X (see safx) more values than levels}

\item{safx}{Values to levels ratio}
}
\value{
Data.frame with (some) variables converted to factors
}
\description{
Load a csv file with read.csv and read_csv
}

