% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{mac_launcher}
\alias{mac_launcher}
\title{Create a launcher and updater for Mac (.command)}
\usage{
mac_launcher(
  app = c("radiant", "radiant.data", "radiant.design", "radiant.basics", "radiant.model",
    "radiant.multivariate"),
  port = 4444,
  pdir = Sys.getenv("HOME")
)
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("radiant", "radiant.data", etc.). Default is "radiant"}

\item{port}{Port to use for shiny::runApp (e.g, port = 4444)}

\item{pdir}{Project directory to use. Default is Sys.getenv("HOME")}
}
\description{
Create a launcher and updater for Mac (.command)
}
\details{
On Mac a file named 'radiant.command' and one named 'update_radiant.command' will be put on the desktop. Double-click the file to launch the specified Radiant app or update Radiant to the latest version
}
\examples{
\dontrun{
radiant::mac_launcher()
}

}
