% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{mac_launcher}
\alias{mac_launcher}
\title{Create a launcher and updater for Mac (.command)}
\usage{
mac_launcher(app = c("radiant", "radiant.data", "radiant.design",
  "radiant.basics", "radiant.model", "radiant.multivariate"))
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("analytics", "marketing", "quant", or "base"). Default is "analytics"}
}
\description{
Create a launcher and updater for Mac (.command)
}
\details{
On Mac a file named 'radiant.command' and one named 'update_radiant.command' will be put on the desktop. Double-click the file to launch the specified Radiant app or update Radiant to the latest version
}

