% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{mac_launcher}
\alias{mac_launcher}
\title{Make a .command launcher for Mac on the desktop}
\usage{
mac_launcher(app = c("marketing", "quant", "base"))
}
\arguments{
\item{app}{App to install ('marketing', 'quant', or 'base'). Default is 'marketing'}
}
\value{
On a mac a file named 'radiant.command' will be put on the desktop. Double-click the file to launch the Radiant app selected
}
\description{
Make a .command launcher for Mac on the desktop
}

