% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{mac_launcher}
\alias{mac_launcher}
\title{Create a launcher for Windows (.command)}
\usage{
mac_launcher(app = c("marketing", "quant", "base"))
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("marketing", "quant", or "base"). Default is "marketing"}
}
\value{
On Mac a file named `radiant.command` will be put on the desktop. Double-click the file to launch the specified Radiant app
}
\description{
Create a launcher for Windows (.command)
}

