% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pivotr.R
\name{make_dt}
\alias{make_dt}
\title{Make a pivot tabel in DT}
\usage{
make_dt(pvt, color_bar = FALSE, perc = FALSE)
}
\arguments{
\item{pvt}{Return value from \code{\link{pivotr}}}

\item{color_bar}{Show color bar in tabel (TRUE, FALSE)}

\item{perc}{Show percentage (e.g., when normalizing)}
}
\description{
Make a pivot tabel in DT
}
\details{
See \url{http://vnijs.github.io/radiant/base/pivotr.html} for an example in Radiant
}
\examples{
pivotr("diamonds", cvars = "cut") \%>\% make_dt
pivotr("diamonds", cvars = c("cut","clarity")) \%>\% make_dt(color_bar = TRUE)
}
\seealso{
\code{\link{pivotr}} to create the pivot-table using dplyr

\code{\link{summary.pivotr}} to print a plain table
}

