% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr.R
\name{make_dt}
\alias{make_dt}
\title{Make a pivot tabel in DT}
\usage{
make_dt(pvt, format = "none", perc = FALSE, dec = 3, search = "",
  searchCols = NULL, order = NULL)
}
\arguments{
\item{pvt}{Return value from \code{\link{pivotr}}}

\item{format}{Show Color bar ("color_bar"),  Heat map ("heat"), or None ("none")}

\item{perc}{Display numbers as percentages (TRUE or FALSE)}

\item{dec}{Number of decimals to show}

\item{search}{Global search. Used to save and restore state}

\item{searchCols}{Column search and filter. Used to save and restore state}

\item{order}{Column sorting. Used to save and restore state}
}
\description{
Make a pivot tabel in DT
}
\details{
See \url{http://vnijs.github.io/radiant/base/pivotr.html} for an example in Radiant
}
\examples{
pivotr("diamonds", cvars = "cut") \%>\% make_dt
pivotr("diamonds", cvars = c("cut","clarity")) \%>\% make_dt(format = "color_bar")
ret <-  pivotr("diamonds", cvars = c("cut","clarity"), normalize = "total") \%>\%
   make_dt(format = "color_bar", perc = TRUE)

}
\seealso{
\code{\link{pivotr}} to create the pivot-table using dplyr

\code{\link{summary.pivotr}} to print a plain text table
}

