% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{make_expl}
\alias{make_expl}
\title{Make a tabel of summary statistics in DT}
\usage{
make_expl(expl, top = "fun", dec = 3, search = "", searchCols = NULL,
  order = NULL)
}
\arguments{
\item{expl}{Return value from \code{\link{explore}}}

\item{top}{The variable (type) to display at the top of the table ("fun" for Function, "var" for Variable, and "byvar" for Group by}

\item{dec}{Number of decimals to show}

\item{search}{Global search. Used to save and restore state}

\item{searchCols}{Column search and filter. Used to save and restore state}

\item{order}{Column sorting. Used to save and restore state}
}
\description{
Make a tabel of summary statistics in DT
}
\details{
See \url{http://vnijs.github.io/radiant/base/explore.html} for an example in Radiant
}
\examples{
tab <- explore("diamonds", "price:x") \%>\% make_expl
tab <- explore("diamonds", "price", byvar = "cut", fun = c("length", "skew")) \%>\%
  make_expl(top = "byvar")

}
\seealso{
\code{\link{pivotr}} to create the pivot-table using dplyr

\code{\link{summary.pivotr}} to print a plain text table
}

