% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ann.R
\name{plot.ann}
\alias{plot.ann}
\title{Plot method for the ann function}
\usage{
\method{plot}{ann}(x, plots = "", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{ann}}}

\item{plots}{Plots to produce for the specified ANN model. Use "" to avoid showing any plots (default). "imp" shows variable importance. "net" shows the network}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the ann function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/ann.html} for an example in Radiant
}
\examples{
result <- ann("titanic", "survived", c("pclass","sex"), lev = "Yes")
plot(result, plots = c("imp","net"))
}
\seealso{
\code{\link{ann}} to generate results

\code{\link{plot.ann}} to plot results

\code{\link{predict.ann}} to generate predictions
}

