% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_means.R
\name{plot.compare_means}
\alias{plot.compare_means}
\title{Plot method for output from compare_means}
\usage{
\method{plot}{compare_means}(result)
}
\description{
Plot method for output from compare_means
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
result <- compare_means("diamonds","cut","price", cm_plots = "box")
plot(result)
}
\seealso{
\code{\link{compare_means}} to calculate results

\code{\link{summary.compare_means}} to summarize results
}

