% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means.R
\name{plot.compare_means}
\alias{plot.compare_means}
\title{Plot method for the compare_means function}
\usage{
\method{plot}{compare_means}(x, plots = "scatter", shiny = FALSE,
  custom = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{compare_means}}}

\item{plots}{One or more plots ("bar", "density", "box", or "scatter")}

\item{shiny}{Did the function call originate inside a shiny app}

\item{custom}{Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This opion can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{http://docs.ggplot2.org/} for options.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the compare_means function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
result <- compare_means("diamonds","cut","price")
plot(result, plots = c("bar","density"))

}
\seealso{
\code{\link{compare_means}} to calculate results

\code{\link{summary.compare_means}} to summarize results
}

