% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_means.R
\name{plot.compare_means}
\alias{plot.compare_means}
\title{Plot method for the compare_means function}
\usage{
\method{plot}{compare_means}(x, cm_plots = "bar", ...)
}
\arguments{
\item{x}{Return value from \code{\link{compare_means}}}

\item{cm_plots}{One or more plots ("bar", "box", or "density")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the compare_means function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/compare_means.html} for an example in Radiant
}
\examples{
result <- compare_means("diamonds","cut","price")
plot(result, cm_plots = c("bar","density"))
}
\seealso{
\code{\link{compare_means}} to calculate results

\code{\link{summary.compare_means}} to summarize results
}

