% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_props.R
\name{plot.compare_props}
\alias{plot.compare_props}
\title{Plot method for the compare_props function}
\usage{
\method{plot}{compare_props}(x, plots = "bar", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{compare_props}}}

\item{plots}{One or more plots of proportions ("bar" or "dodge")}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the compare_props function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/compare_props.html} for an example in Radiant
}
\examples{
result <- compare_props("titanic", "pclass", "survived")
plot(result, plots = c("bar","dodge"))

}
\seealso{
\code{\link{compare_props}} to calculate results

\code{\link{summary.compare_props}} to summarize results
}

