% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint.R
\name{plot.conjoint}
\alias{plot.conjoint}
\title{Plot for conjoint analysis}
\usage{
\method{plot}{conjoint}(result)
}
\description{
Plot for conjoint analysis
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/conjoint.html} for an example in Radiant
}

