% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot.correlation}
\alias{plot.correlation}
\title{Plot method for the correlation function}
\usage{
\method{plot}{correlation}(x, ...)
}
\arguments{
\item{x}{Return value from \code{\link{correlation}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for the correlation function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/correlation.html} for an example in Radiant
}
\examples{
result <- correlation("diamonds",c("price","carat","clarity"))
plot(result)
}
\seealso{
\code{\link{correlation}} to calculate results

\code{\link{summary.correlation}} to summarize results
}

