% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm.R
\name{plot.glm_reg}
\alias{plot.glm_reg}
\title{Plot method for glm_reg}
\usage{
\method{plot}{glm_reg}(result)
}
\arguments{
\item{result}{The selected plot type}
}
\description{
Plot method for glm_reg
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/glm.html} for an example in Radiant
}
\examples{
result <- glm_reg("titanic", "survived", c("pclass","sex"), glm_levels = "Yes", glm_plots = "coef")
plot(result)
}
\seealso{
\code{\link{glm_reg}} to generate results

\code{\link{plot.glm_reg}} to plot results
}

