% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mds.R
\name{plot.mds}
\alias{plot.mds}
\title{Plot method for the mds function}
\usage{
\method{plot}{mds}(x, mds_rev_dim = "", mds_fontsz = 1.3, ...)
}
\arguments{
\item{x}{Return value from \code{\link{mds}}}

\item{mds_rev_dim}{Flip the axes in plots}

\item{mds_fontsz}{Font size to use in plots}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the mds function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/mds.html} for an example in Radiant
}
\examples{
result <- mds("city","from","to","distance")
plot(result)
plot(result, mds_rev_dim = 1:2)
plot(result, mds_rev_dim = 1:2, mds_fontsz = 2)
}
\seealso{
\code{\link{mds}} to calculate results

\code{\link{summary.mds}} to plot results
}

