% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pivotr.R
\name{plot.pivotr}
\alias{plot.pivotr}
\title{Plot method for the pivotr function}
\usage{
\method{plot}{pivotr}(x, type = "dodge", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{pivotr}}}

\item{type}{Plot type to use ("fill" or "dodge" (default))}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the pivotr function
}
\details{
See \url{http://vnijs.github.io/radiant/base/pivotr} for an example in Radiant
}
\seealso{
\code{\link{pivotr}} to generate summaries

\code{\link{summary.pivotr}} to show summaries
}

