% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre_factor.R
\name{plot.pre_factor}
\alias{plot.pre_factor}
\title{Plot method for the pre_factor function}
\usage{
\method{plot}{pre_factor}(x, ...)
}
\arguments{
\item{x}{Return value from \code{\link{pre_factor}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the pre_factor function
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/pre_factor.html} for an example in Radiant
}
\examples{
result <- pre_factor("diamonds",c("price","carat","table"))
plot(result)
}
\seealso{
\code{\link{pre_factor}} to calculate results

\code{\link{summary.pre_factor}} to summarize results
}

