% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_disc}
\alias{plot.prob_disc}
\title{Plot method for the probability calculator function (discrete)}
\usage{
\method{plot}{prob_disc}(x, type = "values", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_disc}}}

\item{type}{Probabilities or values}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator function (discrete)
}
\details{
See \url{http://vnijs.github.io/radiant/quant/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_disc(v = "5 6 7 8 9 10 11 ", p = ".1 .2 .3 .15 .1 .1 .05", pub = 0.95)
plot(result, type = "probs")

}

