% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulater.R
\name{plot.repeater}
\alias{plot.repeater}
\title{Plot repeated simulation}
\usage{
\method{plot}{repeater}(x, sum_vars = "", byvar = "sim", fun = "sum_rm",
  form = "", shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{repeater}}}

\item{sum_vars}{(Numerical) variables to summaries}

\item{byvar}{Variable(s) to group data by before summarizing}

\item{fun}{Functions to use for summarizing}

\item{form}{A string with the formula to evaluate (e.g., "profit = demand * (price - cost)")}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot repeated simulation
}

