% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_mean.R
\name{plot.single_mean}
\alias{plot.single_mean}
\title{Plot results from the single_mean function. This is a method of class single_mean and can be called as plot or plot.single_mean}
\usage{
\method{plot}{single_mean}(result)
}
\arguments{
\item{result}{Return value from \code{\link{single_mean}}}
}
\description{
Plot results from the single_mean function. This is a method of class single_mean and can be called as plot or plot.single_mean
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_mean.html} for an example in Radiant
}
\examples{
result <- single_mean("diamonds","price")
plot(result)
}
\seealso{
\code{\link{single_mean}} to generate the result

\code{\link{summary.single_mean}} to summarize results
}

