% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{print.gtable}
\alias{print.gtable}
\title{Print/draw method for grobs produced by gridExtra}
\usage{
\method{print}{gtable}(x, ...)
}
\arguments{
\item{x}{a gtable object}

\item{...}{further arguments passed to or from other methods}
}
\value{
A plot
}
\description{
Print/draw method for grobs produced by gridExtra
}
\details{
Print method for ggplot grobs created using arrangeGrob. Code is based on \url{https://github.com/baptiste/gridextra/blob/master/inst/testing/shiny.R}
}

