% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant}
\alias{radiant}
\alias{radiant-package}
\alias{radiant}
\title{radiant}
\usage{
radiant(run = "browser")
}
\arguments{
\item{run}{Run radiant in an external browser ("browser") or in the Rstudio viewer ("viewer")}
}
\description{
radiant

Launch radiant in default browser or Rstudio Viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant::radiant()
}
}
