% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant}
\alias{radiant}
\alias{radiant-package}
\title{radiant}
\usage{
radiant()
}
\description{
radiant

Launch Radiant in the default browser
}
\details{
See \url{http://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant()
}
}

