% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant}
\alias{radiant}
\title{radiant}
\usage{
radiant(state, ...)
}
\arguments{
\item{state}{Path to state file to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch radiant in the default web browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant()
radiant("https://github.com/radiant-rstats/docs/raw/gh-pages/examples/demo-dvd-rnd.state.rda")
}
}
