% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant}
\alias{radiant}
\title{Launch the Radiant app in a browser}
\usage{
radiant(app = c("marketing", "quant"))
}
\description{
Launch the Radiant app in a browser
}
\details{
See \url{http://mostly-harmless.github.io/radiant} for documentation and tutorials
}

