% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant}
\alias{radiant}
\alias{radiant-package}
\title{radiant}
\usage{
radiant(app = c("analytics", "marketing", "quant", "base"))
}
\arguments{
\item{app}{Choose the app to run. One of "base", "quant", "analytics", "marketing". "analytics" is the default}
}
\description{
radiant

Launch Radiant in the default browser
}
\details{
See \url{http://vnijs.github.io/radiant} for documentation and tutorials
}
\examples{
if (interactive()) {
  radiant("base")
  radiant("quant")
  radiant("marketing")
  radiant("analytics")
}
}

