% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant}
\alias{radiant}
\alias{radiant-package}
\title{radiant}
\usage{
radiant(app = c("marketing", "quant", "base"))
}
\arguments{
\item{app}{Choose the app to run. Either "base", "quant", or "marketing". "marketing" is the default}
}
\description{
radiant

Launch Radiant in the default browser
}
\details{
See \url{http://mostly-harmless.github.io/radiant} for documentation and tutorials
}

