% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant_viewer}
\alias{radiant_viewer}
\title{Launch radiant in the Rstudio viewer}
\usage{
radiant_viewer()
}
\description{
Launch radiant in the Rstudio viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant_viewer()
}
}
