% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{rndnames}
\alias{rndnames}
\title{A list of 100 random names generated by listofrandomnames.com
Description provided in attr(rndnames,"description")}
\format{A data frame with 100 rows and 2 variables}
\usage{
data(rndnames)
}
\description{
A list of 100 random names generated by listofrandomnames.com
Description provided in attr(rndnames,"description")
}
\keyword{datasets}

