% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_comp.R
\name{sample_size_comp}
\alias{sample_size_comp}
\title{Sample size calculation for comparisons}
\usage{
sample_size_comp(type, n = NULL, p1 = 0.5, p2 = NULL, delta = NULL,
  sd = 1, conf_lev = 0.95, power = NULL, alternative = "two.sided")
}
\arguments{
\item{type}{Choose "mean" or "proportion"}

\item{power}{Power}

\item{conf}{Confidence level}

\item{diff}{Detectible difference}

\item{prop}{Initial proportion}

\item{alt}{Two or one sided test}
}
\value{
A list of variables defined in sample_size_comp as an object of class sample_size_comp
}
\description{
Sample size calculation for comparisons
}
\details{
See \url{http://vnijs.github.io/radiant/quant/sample_size_comp.html} for an example in Radiant
}
\seealso{
\code{\link{summary.sample_size_comp}} to summarize results
}

