% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/full_factor.R
\name{save_factors}
\alias{save_factors}
\title{Save factor scores to active dataset}
\usage{
save_factors(object)
}
\arguments{
\item{object}{Return value from \code{\link{full_factor}}}
}
\description{
Save factor scores to active dataset
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/full_factor.html} for an example in Radiant
}
\examples{
\dontrun{
result <- full_factor("diamonds",c("price","carat","table"))
save_factors(result)
head(dat)
}
}

