% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{save_membership}
\alias{save_membership}
\title{Add a cluster membership variable to the active dataset}
\usage{
save_membership(object)
}
\arguments{
\item{object}{Return value from \code{\link{kmeans_clus}}}
}
\description{
Add a cluster membership variable to the active dataset
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/kmeans_clus.html} for an example in Radiant
}
\examples{
\dontrun{
result <- kmeans_clus("shopping", km_vars = c("v1:v6"))
save_membership(result)
}
}
\seealso{
\code{\link{kmeans_clus}} to generate results

\code{\link{summary.kmeans_clus}} to summarize results

\code{\link{plot.kmeans_clus}} to plot results

\code{\link{save_kmeans}} to save the table of cluster means to a csv file
}

