% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{sd_rm}
\alias{sd_rm}
\title{Standard deviation with na.rm = TRUE}
\usage{
sd_rm(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{Remove NAs (TRUE or FALSE)}
}
\value{
Standard deviation
}
\description{
Standard deviation with na.rm = TRUE
}
\examples{
sd_rm(rnorm(100))

}

