% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{set_class}
\alias{set_class}
\title{Alias used to set the class for output from an analysis function in R/ (e.g., single_mean)}
\usage{
set_class()
}
\description{
Alias used to set the class for output from an analysis function in R/ (e.g., single_mean)
}
\examples{
foo <- function(x) x^2 \%>\% set_class(c("foo", class(.)))
}

