% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{set_class}
\alias{set_class}
\title{Alias used to set the class for analysis function return}
\usage{
set_class()
}
\description{
Alias used to set the class for analysis function return
}
\examples{
foo <- function(x) x^2 \%>\% set_class(c("foo", class(.)))

}

