% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{sig_stars}
\alias{sig_stars}
\title{Add stars '***' to a data.frame (from broom's `tidy` function) based on p.values}
\usage{
sig_stars(pval)
}
\arguments{
\item{p.value}{Vector of p.values from an analysis}
}
\value{
A list with all variables defined in the function
}
\description{
Add stars '***' to a data.frame (from broom's `tidy` function) based on p.values
}
\examples{
sig_stars(c(.0009, .049, .009, .4, .09))
}

