% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{sshh}
\alias{sshh}
\title{Hide warnings and messages from app user (e.g., ggplot2). Adapted from http://www.onthelambda.com/2014/09/17/fun-with-rprofile-and-customizing-r-startup/}
\usage{
sshh(...)
}
\arguments{
\item{...}{Inputs to keep quite}
}
\description{
Hide warnings and messages from app user (e.g., ggplot2). Adapted from http://www.onthelambda.com/2014/09/17/fun-with-rprofile-and-customizing-r-startup/
}
\examples{
sshh( library(dplyr) )
}

