% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{sshhr}
\alias{sshhr}
\title{Hide warnings and messages and return result}
\usage{
sshhr(...)
}
\arguments{
\item{...}{Inputs to keep quite}
}
\description{
Hide warnings and messages and return result
}
\details{
Adapted from \url{http://www.onthelambda.com/2014/09/17/fun-with-rprofile-and-customizing-r-startup/}
}
\examples{
sshhr( library(dplyr) )

}

