% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.R
\name{store_ann}
\alias{store_ann}
\title{Store predicted values generated in the ann function}
\usage{
store_ann(pred, data, name = "predict_ann")
}
\arguments{
\item{pred}{Return value from predict.nnet}

\item{data}{Dataset name}

\item{name}{Variable name assigned to the predicted values}
}
\description{
Store predicted values generated in the ann function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/ann.html} for an example in Radiant
}

