% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{store_glm}
\alias{store_glm}
\title{Store residuals or predicted values generated in the glm_reg function}
\usage{
store_glm(object, data = object$dataset, type = "residuals",
  name = paste0(type, "_glm"))
}
\arguments{
\item{object}{Return value from \code{\link{glm_reg}} or \code{\link{predict.glm_reg}}}

\item{data}{Dataset name}

\item{type}{Residuals ("residuals") or predictions ("predictions"). For predictions the dataset name must be provided}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Store residuals or predicted values generated in the glm_reg function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}
\examples{
\donttest{
result <- glm_reg("titanic", "survived", "pclass", lev = "Yes")
store_glm(result)
}
}

