% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{store_reg}
\alias{store_reg}
\title{Store residuals or predicted values generated in the regression function}
\usage{
store_reg(object, data = object$dataset, type = "residuals",
  name = paste0(type, "_reg"))
}
\arguments{
\item{object}{Return value from \code{\link{regression}} or \code{\link{predict.regression}}}

\item{data}{Dataset name}

\item{type}{Residuals ("residuals") or predictions ("predictions"). For predictions the dataset name must be provided}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Store residuals or predicted values generated in the regression function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
\donttest{
result <- regression("diamonds", "price", c("carat","clarity"))
store_reg(result)
}
}

