% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_props.R
\name{summary.compare_props}
\alias{summary.compare_props}
\title{Summary method for the compare_props function}
\usage{
\method{summary}{compare_props}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{compare_props}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the compare_props function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/compare_props.html} for an example in Radiant
}
\examples{
result <- compare_props("titanic", "pclass", "survived")
summary(result)
}
\seealso{
\code{\link{compare_props}} to calculate results

\code{\link{plot.compare_props}} to plot results
}

