% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint.R
\name{summary.conjoint}
\alias{summary.conjoint}
\title{Summary for conjoint analysis}
\usage{
\method{summary}{conjoint}(result)
}
\description{
Summary for conjoint analysis
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/conjoint.html} for an example in Radiant
}

