% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint.R
\name{summary.conjoint}
\alias{summary.conjoint}
\title{Summary method for conjoint output}
\usage{
\method{summary}{conjoint}(result)
}
\description{
Summary method for conjoint output
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/conjoint.html} for an example in Radiant
}
\examples{
result <- conjoint(dataset = "mp3", ca_var1 = "Rating", ca_var2 = c("Memory", "Radio", "Size", "Price", "Shape"), ca_scale_plot = TRUE)
summary(result)
}
\seealso{
\code{\link{conjoint}} to generate results

\code{\link{plot.conjoint}} to plot results
}

