% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conjoint_profiles.R
\name{summary.conjoint_profiles}
\alias{summary.conjoint_profiles}
\title{Summary method for the conjoint_profiles function}
\usage{
\method{summary}{conjoint_profiles}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{conjoint_profiles}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary method for the conjoint_profiles function
}
\details{
See \url{http://vnijs.github.io/radiant/marketing/conjoint_profiles.html} for an example in Radiant
}
\examples{
\donttest{
cp <<- c("price = c('$10','$13','$16')", "sight = c('Staggered','Not Staggered')",
       "comfort = c('Average no cupholder','Average cupholder','Large cupholder')",
       "audio.visual = c('Small plain','Large plain','Large digital')",
       "food = c('No food','Hot dogs and popcorn','Gourmet food')")
result <- conjoint_profiles("cp")
summary(result)
rm(cp, envir = .GlobalEnv)
}
}
\seealso{
\code{\link{conjoint_profiles}} to calculate results
}

