% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conjoint_profiles.R
\name{summary.conjoint_profiles}
\alias{summary.conjoint_profiles}
\title{Summarize method for the conjoint_profiles function}
\usage{
\method{summary}{conjoint_profiles}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{conjoint_profiles}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarize method for the conjoint_profiles function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/conjoint_profiles.html} for an example in Radiant
}
\examples{
ca_prof <- readLines(system.file("examples/profiles-movie.txt", package='radiant'))
result <- conjoint_profiles("ca_prof")
summary(result)
}
\seealso{
\code{\link{conjoint_profiles}} to calculate results
}

