% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation.R
\name{summary.correlation}
\alias{summary.correlation}
\title{Summarize method for the correlation function}
\usage{
\method{summary}{correlation}(object, cor_cutoff = 0, ...)
}
\arguments{
\item{object}{Return value from \code{\link{correlation}}}

\item{cor_cutoff}{Show only corrlations larger than the cutoff in absolute value. Default is a cutoff of 0}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarize method for the correlation function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/correlation.html} for an example in Radiant
}
\examples{
result <- correlation("diamonds",c("price","carat","clarity"))
summary(result, cor_cutoff = .3)
}
\seealso{
\code{\link{correlation}} to calculate results

\code{\link{plot.correlation}} to plot results
}

