% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cross_tabs.R
\name{summary.cross_tabs}
\alias{summary.cross_tabs}
\title{Summary method for the cross_tabs function}
\usage{
\method{summary}{cross_tabs}(result)
}
\description{
Summary method for the cross_tabs function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/cross_tabs.html} for an example in Radiant
}
\examples{
result <- cross_tabs("newspaper", "Income", "Newspaper")
summary(result)
}
\seealso{
\code{\link{cross_tabs}} to calculate results

\code{\link{plot.cross_tabs}} to plot results
}

