% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm_reg.R
\name{summary.glm_reg}
\alias{summary.glm_reg}
\title{Summary method for the glm_reg function}
\usage{
\method{summary}{glm_reg}(object, sum_check = "", conf_lev = 0.95,
  test_var = "", ...)
}
\arguments{
\item{object}{Return value from \code{\link{glm_reg}}}

\item{sum_check}{Optional output or estimation parameters. "rsme" to show the root mean squared error. "sumsquares" to show the sum of squares table. "vif" to show multicollinearity diagnostics. "confint" to show coefficient confidence interval estimates.}

\item{conf_lev}{Confidence level to use for coefficient and odds confidence intervals (.95 is the default)}

\item{test_var}{Variables to evaluate in model comparison (i.e., a competing models Chi-squared test)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the glm_reg function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}
\examples{
result <- glm_reg("titanic", "survived", "pclass", lev = "Yes")
summary(result, test_var = "pclass")
res <- glm_reg("titanic", "survived", c("pclass","sex"), int_var="pclass:sex", lev="Yes")
summary(res, sum_check = c("vif","confint","odds"))
titanic \%>\% glm_reg("survived", c("pclass","sex","age"), lev = "Yes") \%>\% summary("vif")
}
\seealso{
\code{\link{glm_reg}} to generate the results

\code{\link{plot.glm_reg}} to plot the results

\code{\link{predict.glm_reg}} to generate predictions

\code{\link{plot.glm_predict}} to plot prediction output
}

