% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pivotr.R
\name{summary.pivotr}
\alias{summary.pivotr}
\title{Summary method for pivotr}
\usage{
\method{summary}{pivotr}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{pivotr}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for pivotr
}
\details{
See \url{http://vnijs.github.io/radiant/base/pivotr.html} for an example in Radiant
}
\examples{
pivotr("diamonds", cvars = "cut") \%>\% summary
pivotr("diamonds", cvars = "cut") \%>\% summary
pivotr("diamonds", cvars = "cut:clarity", nvar = "price") \%>\% summary
}
\seealso{
\code{\link{pivotr}} to create the pivot-table using dplyr
}

