% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre_factor.R
\name{summary.pre_factor}
\alias{summary.pre_factor}
\title{Summarize method for the pre_factor function}
\usage{
\method{summary}{pre_factor}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{pre_factor}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summarize method for the pre_factor function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/marketing/pre_factor.html} for an example in Radiant
}
\examples{
result <- pre_factor("diamonds",c("price","carat","table"))
summary(result)
result <- pre_factor("computer","HighEnd:Business")
summary(result)
}
\seealso{
\code{\link{pre_factor}} to calculate results

\code{\link{plot.pre_factor}} to plot results
}

