% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{summary.regression}
\alias{summary.regression}
\title{Summary method for the regression function}
\usage{
\method{summary}{regression}(object, sum_check = "", conf_lev = 0.95,
  test_var = "", ...)
}
\arguments{
\item{object}{Return value from \code{\link{regression}}}

\item{sum_check}{Optional output or estimation parameters. "rsme" to show the root mean squared error. "sumsquares" to show the sum of squares table. "vif" to show multicollinearity diagnostics. "confint" to show coefficient confidence interval estimates.}

\item{conf_lev}{Confidence level used to estimate confidence intervals (.95 is the default)}

\item{test_var}{Variables to evaluate in model comparison (i.e., a competing models F-test)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the regression function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
result <- regression("diamonds", "price", c("carat","clarity"))
summary(result, sum_check = c("rmse","sumsquares","vif","confint"), test_var = "clarity")
result <- regression("shopping", "v1", c("v2","v3"))
summary(result, test_var = "v2")
shopping \%>\% regression("v1", "v2:v6") \%>\% summary
}
\seealso{
\code{\link{regression}} to generate the results

\code{\link{plot.regression}} to plot results

\code{\link{predict.regression}} to generate predictions
}

