% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{summary.sampling}
\alias{summary.sampling}
\title{Summary method for the sampling function}
\usage{
\method{summary}{sampling}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{sampling}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the sampling function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/sampling} for an example in Radiant
}
\examples{
result <- sampling("rndnames","Names",10)
summary(result)
rndnames \%>\% sampling("Names",10) \%>\% summary
}
\seealso{
\code{\link{sampling}} to generate the results
}

