% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_mean.R
\name{summary.single_mean}
\alias{summary.single_mean}
\title{Summary method for single_mean}
\usage{
\method{summary}{single_mean}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{single_mean}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for single_mean
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_mean.html} for an example in Radiant
}
\examples{
result <- single_mean("diamonds","price")
summary(result)
}
\seealso{
\code{\link{single_mean}} to generate the results

\code{\link{plot.single_mean}} to plot results
}

