% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_prop.R
\name{summary.single_prop}
\alias{summary.single_prop}
\title{Summary method for the single_prop function}
\usage{
\method{summary}{single_prop}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{single_prop}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the single_prop function
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/single_prop.html} for an example in Radiant
}
\examples{
result <- single_prop("diamonds","clarity", sp_levels = "IF", sp_comp_value = 0.05)
summary(result)
}
\seealso{
\code{\link{single_prop}} to generate the results

\code{\link{plot.single_prop}} to plot the results
}

