% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{var_check}
\alias{var_check}
\title{Check if main effects for all interaction effects are included in the model
If ':' is used to select a range _indep_var_ is updated}
\usage{
var_check(iv, cn, intv = "")
}
\arguments{
\item{iv}{List of independent variables provided to _regression_ or _glm_}

\item{cn}{Column names for all independent variables in _dat_}

\item{intv}{Interaction terms specified}
}
\value{
'vars' is a vector of right-hand side variables, possibly with interactions, 'iv' is the list of independent variables, and intv are interaction terms
}
\description{
Check if main effects for all interaction effects are included in the model
If ':' is used to select a range _indep_var_ is updated
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
var_check("a:d", c("a","b","c","d"))
var_check(c("a", "b"), c("a", "b"), "a:c")
}

