% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression.R
\name{var_check}
\alias{var_check}
\title{Check if main effects for all interaction effects are included in the model
If ':' is used to select a range _indep_var_ is updated}
\usage{
var_check(indep_var, cn, int_var = "")
}
\arguments{
\item{indep_var}{List of independent variables provided to _regression_ or _glm_}

\item{cn}{Column names for all independent variables in _dat_}

\item{int_var}{Interaction terms specified}
}
\value{
'vars' is a vector of right-hand side variables, possibly with interactions, 'indep_var' is the list of indepdent variables, and int_var are interaction terms
}
\description{
Check if main effects for all interaction effects are included in the model
If ':' is used to select a range _indep_var_ is updated
}
\details{
See \url{http://mostly-harmless.github.io/radiant/quant/regression.html} for an example in Radiant
}
\examples{
var_check("a:d", c("a","b","c","d"))
var_check(c("a","b"), c("a","b"), "a:c")
}

