% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{viewdata}
\alias{viewdata}
\title{View data}
\usage{
viewdata(dataset, vars = "", filt = "")
}
\arguments{
\item{dataset}{Name of the dataframe to change}

\item{vars}{Variables to so (default is all)}

\item{filt}{Filter to apply to the specified dataset. For example "price > 10000" if dataset is "diamonds" (default is "")}
}
\description{
View data
}
\details{
View, search, sort, etc. your data
}
\examples{
if (interactive()) {
viewdata(mtcars)
viewdata("mtcars")
mtcars \%>\% viewdata
}
}

