% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{win_launcher}
\alias{win_launcher}
\title{Create a launcher and updater for Windows (.bat)}
\usage{
win_launcher(
  app = c("radiant", "radiant.data", "radiant.design", "radiant.basics", "radiant.model",
    "radiant.multivariate"),
  port = 4444,
  pdir = Sys.getenv("HOME")
)
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("radiant", "radiant.data", etc.). Default is "radiant"}

\item{port}{Port to use for shiny::runApp (e.g, port = 4444)}

\item{pdir}{Project directory to use. Default is Sys.getenv("HOME")}
}
\description{
Create a launcher and updater for Windows (.bat)
}
\details{
On Windows a file named 'radiant.bat' and one named 'update_radiant.bat' will be put on the desktop. Double-click the file to launch the specified Radiant app or update Radiant to the latest version
}
\examples{
\dontrun{
radiant::win_launcher()
}

}
