% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radiant.R
\name{win_launcher}
\alias{win_launcher}
\title{Create a launcher for Windows (.bat)}
\usage{
win_launcher(app = c("marketing", "quant", "base"))
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("marketing", "quant", or "base"). Default is "marketing"}
}
\description{
Create a launcher for Windows (.bat)
}
\details{
On Windows a file named `radiant.bat` will be put on the desktop. Double-click the file to launch the specified Radiant app
}
\examples{
if (interactive()) {
  if(Sys.info()["sysname"] != "Windows") {
    win_launcher()
    fn <- paste0(Sys.getenv("USERPROFILE") ,"/Desktop/radiant.bat")
    if(!file.exists(fn))
       stop("Windows launcher not created")
    else
      unlink(fn)
  }
}
}

