% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{xtile}
\alias{xtile}
\title{Create a quintile (or decile) index}
\usage{
xtile(x, n, rev = FALSE)
}
\arguments{
\item{x}{Numeric variable}

\item{n}{number of bins to create}

\item{rev}{Reverse the order of the xtiles}
}
\description{
Create a quintile (or decile) index
}
\details{
Same as stata
}
\examples{
xtile(1:10,5)
xtile(1:10,5, rev = TRUE)

}

